% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{AddElevationData}
\alias{AddElevationData}
\title{Adds elevation data to locations dataframe}
\usage{
AddElevationData(df, long, lat)
}
\arguments{
\item{df}{dataframe of location data}

\item{long}{column name of longitude data, default = lat_utm}

\item{lat}{column name of latitude data, default = long_utm}
}
\value{
The original locations df with additional landscape data columns.
}
\description{
Adds the cell values from an elevation rasters to an input locations and
  calculates the 'agl' (above ground level) of the locations.
}
\details{
The locations dataset and elevation data must have the same CRS.
}
