% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{AddFirstLastDistance}
\alias{AddFirstLastDistance}
\title{Adds distance to first and last locations for each day}
\usage{
AddFirstLastDistance(df, id, datetime, tz)
}
\arguments{
\item{df}{dataframe with locations}

\item{by}{column name of unique identifier, default is "id"}

\item{datetime}{column name of datetime in POSIXct format or as a character
in the format (\%Y/\%m/\%d \%H:\%M)}

\item{tz}{return to this timezome. Default "Etc/GMT+5".}
}
\value{
dataframe with "datetime", "first", "last"
}
\description{
Adds columns with distance to first and last location of each day
}
\details{
Caution with use of tz, may cause problems
}
