% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{AddLandscapeValues}
\alias{AddLandscapeValues}
\title{Adds landscape data to dataframe}
\usage{
AddLandscapeValues(df, raster_stack, long, lat, clean_up)
}
\arguments{
\item{df}{dataframe of location data}

\item{raster_stack}{a RasterStack of landscape data}

\item{long}{column name of longitude data, default = lat_utm}

\item{lat}{column name of latitude data, default = long_utm}

\item{clean_up}{logical, will run a series of steps to clean up the returned
output (e.g. removing "_50mc" from column names and deleting unneeded
landcover "definition" column).}
}
\value{
The original locations df with additional landscape data columns.
}
\description{
Adds the cell values from a list of landscape rasters to an input locations
  dataset, usually a subset of BAEA.
}
\details{
The locations dataset and rasters should have the same CRS,
  otherwise function stops and returns an error message.
}
