% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{AddNestData}
\alias{AddNestData}
\title{Adds nest-related columns to dataframe}
\usage{
AddNestData(df, nests_study, nests_use)
}
\arguments{
\item{df}{dataframe of BAEA location data}

\item{nests_study}{.csv file of study nests. Default is:
"C:/Work/R/Data/BAEA/Nests/Nests_Study.csv"}

\item{nests_use}{.csv file of study nests corresponding use dates. Default
is: "C:/Work/R/Data/BAEA/Nests/Nests_Study_Use_Dates.csv"}
}
\value{
dataframe
}
\description{
Adds nest locations, angle, and distance to each record, based on "date" and
  "id"
}
