% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{AddSolarTimes}
\alias{AddSolarTimes}
\title{Add solar times}
\usage{
AddSolarTimes(df, id, tz)
}
\arguments{
\item{df}{Dataframe}

\item{by}{Column name to use to split, analyze, and merge data. Default is
"id".}

\item{tz}{Column name of timezone. Default is "Etc/GMT+5".}
}
\value{
Dataframe with sunrise, sunset, and solarnoon times
}
\description{
Add sunrise, sunset, and solarnoon to a dataframe of location data
}
\details{
Coordinates in "lat" and "long" must be WGS84, sunrise and
  solarnoon based on first location, sunset based on last location.
}
