% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{AddStepLengthAndAngles}
\alias{AddStepLengthAndAngles}
\title{Add step length and angle to location data}
\usage{
AddStepLengthAndAngles(df, by, datetime, long, lat)
}
\arguments{
\item{df}{dataframe with locations and datetime data}

\item{by}{column name to use to split, analyze, and merge data, default is
"id"}

\item{long}{longitude, must be in UTM, default is "long_utm"}

\item{lat}{latitude, must be in UTM, default is "lat_utm"}
}
\value{
dataframe with "step_length", "abs_angle", and "turn_angle" columns
}
\description{
Calculates step length, absolute angle (N=0), and turn angle between
  successive point locations
}
\details{
Coordinates must be in identically-scaled units (e.g. UTM meters).
  If lat and long are in degrees, project coordinates to UTM with 'rgdal'
  package before running this function.
}
