% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{AddStepTime}
\alias{AddStepTime}
\title{Add step time}
\usage{
AddStepTime(df, by, datetime)
}
\arguments{
\item{df}{Dataframe with locations and datetime data}

\item{by}{Column name to use to split, analyze, and merge data, default is
"id".}

\item{datetime}{Column name of datetime, default is "datetime"}
}
\value{
Dataframe with "step_time" column
}
\description{
Calculates time between successive point locations
}
