% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{AddUTMCoordinates}
\alias{AddUTMCoordinates}
\title{Add UTM coordinates to a dataframe}
\usage{
AddUTMCoordinates(df, long, lat, crs)
}
\arguments{
\item{df}{dataframe of location data}

\item{long}{column name of longitude data, default = "lat"}

\item{lat}{column name of latitude data, default = "long"}

\item{crs}{Projected Coordinate System, default is
"+proj=utm +zone=19 ellps=WGS84"}
}
\value{
The original locations df with "long_utm" and "lat_utm" values.
}
\description{
Adds UTM_lat and UTM_long values to a dataframe with "long", "lat"
  coordinates
}
