% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admins_map.R
\name{admins_map}
\alias{admins_map}
\title{Generate administrative boundaries map}
\usage{
admins_map(countries, add_neighbors = FALSE)
}
\arguments{
\item{countries}{list of countries to map}

\item{add_neighbors}{should the map include the neighbor countries}
}
\value{

}
\description{
Generate administrative boundaries map
}
\examples{
\dontrun{
library(gisr)
admins_map(countries = list("Zambia"))
admins_map(countries = list("Zambia"), add_neighbors = TRUE)
}

}
