% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_locations.R
\name{assess_facilities}
\alias{assess_facilities}
\title{Assess facility geo-location reporting levels}
\usage{
assess_facilities(.data)
}
\arguments{
\item{.data}{Datim organisation units data frame}
}
\description{
Assess facility geo-location reporting levels
}
\examples{
\dontrun{
assess_facilities(df_sites)
df_sites \%>\% assess_facilities()
}
}
