% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{attributes}
\alias{attributes}
\title{Get attributes from feature class}
\usage{
attributes(geodata)
}
\arguments{
\item{geodata}{spatial data}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    attribures as data frame
}\if{html}{\out{</div>}}
}
\description{
Get attributes from feature class
}
\examples{
\donttest{
 library(gisr)

 # Admin level 0 [country] geodata
 adm0 <- get_admin0("Ghana")

 # Extract attrs from geodata
 df <- attributes(geodata = adm0)

 df \%>\% head()
}

}
