% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{CalculateTerrainMetric}
\alias{CalculateTerrainMetric}
\title{Calculate terrain metrics on elevation RasterLayer}
\usage{
CalculateTerrainMetric(elev, size, metric)
}
\arguments{
\item{elev}{RasterLayer, elevation data raster}

\item{size}{numeric, dimensions of analysis window (nrow and ncol). Must be
odd number.}

\item{metric}{character, vector containing one of these options: tpi, tri,
roughness (see Details)}
}
\value{
RasterLayer
}
\description{
Calculate terrain metrics on elevation RasterLayer
}
\details{
metrics based on examples given in terrain() in 'raster' package
}
