% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{CenterXYWithBase}
\alias{CenterXYWithBase}
\title{Centers x,y data based on a 'base' raster}
\usage{
CenterXYWithBase(df, base)
}
\arguments{
\item{df}{dataframe with x and y columns}

\item{base}{raster base layer}
}
\value{
Dataframe with centered x and y (i.e., (x,y))
}
\description{
Centers x and y values of a dataframe into the center of raster cells based
  on the base layer
}
