% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_columns}
\alias{check_columns}
\title{Check columns length}
\usage{
check_columns(spdf)
}
\arguments{
\item{spdf}{input spatial data frame}
}
\value{
list of columns with more than 10 characters
}
\description{
Check columns length
}
