% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean_levels}
\alias{clean_levels}
\title{Clean org hierarchy levels}
\usage{
clean_levels(.df_levels)
}
\value{
df
}
\description{
Clean org hierarchy levels
}
\examples{
\dontrun{
  library(gisr)

  # Get PEPFAR Org Levels
  clean_levels(df_lvls)
 }

}
