% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{cntry_polygons}
\alias{cntry_polygons}
\title{Extract Country Polygons}
\usage{
cntry_polygons(spdf, cntry)
}
\arguments{
\item{spdf}{VcPolygons data as Spatial Data Frame}

\item{cntry}{Country name}
}
\value{
list of spatial data frames
}
\description{
Extract Country Polygons
}
\examples{
\dontrun{

spdf <- gisr::get_vcpolygons(path = glamr::si_path("path_vector"), name = "VcPepfarPolygons.shp")

cntry_polygons(spdf = spdf, name = "Zambia")
}

}
