% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{CompileDownloads}
\alias{CompileDownloads}
\title{Compile CTT downloads}
\usage{
CompileDownloads(units, compile, tz)
}
\arguments{
\item{units}{String, "deployed" or "reserve". Default is "deployed".}

\item{compile}{String, "all" or "recent". Default is "recent"}

\item{tz}{Timezone, default is "Etc/GMT+5".}
}
\value{
Dataframe of compiled files
}
\description{
Compiles the downloaded CTT files
}
\details{
Internal parameters are set specifically for my project's data
}
