% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{ConvertRasterForGGPlot}
\alias{ConvertRasterForGGPlot}
\title{Converts raster for use in ggplot}
\usage{
ConvertRasterForGGPlot(raster, sample_size, na.rm)
}
\arguments{
\item{raster}{Raster file}

\item{sample_size}{integer, raster sample size. Default is 10000.}

\item{na.rm}{logical, remove NA values. Default is TRUE.}
}
\value{
dataframe
}
\description{
A function for converting a Raster file into a df for use in a ggplot
}
\details{
most of script is from rasterVis package
}
