% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{ConvertToVoronoi}
\alias{ConvertToVoronoi}
\title{Converts points to Voronoi map}
\usage{
ConvertToVoronoi(spdf, return_df = TRUE)
}
\arguments{
\item{spdf}{a SpatialPointsDataFrame}

\item{return_df}{logical, whether to return a dataframe. Default is TRUE.}
}
\value{
dataframe or SpatialPolygonDataFrame
}
\description{
Converts a SpatialPointsDataFrame to a Voronoi map as dataframe or a
SpatialPolygonDataFrame
}
\details{
Modified from R-Bloggers - "Making Static/Interactive Voronoi Map"
}
