% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{CreateArcGISandPDFMaps}
\alias{CreateArcGISandPDFMaps}
\title{Create ArcGIS and PDF maps of Bald Eagle data}
\usage{
CreateArcGISandPDFMaps(df)
}
\arguments{
\item{df}{dataframe with locations}
}
\value{
overwrites files in "C:/Work/Python" folders
}
\description{
A wrapper function for creating an ArcGIS map and a series of PDFs
}
\details{
internal parameters of Python script specific to my BAEA GPS data.
  Identical to CreateArcGISMaps(), except it also creates PDFs.
}
