% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{CreateArcGISMaps}
\alias{CreateArcGISMaps}
\title{Create ArcGIS maps of Bald Eagle data}
\usage{
CreateArcGISMaps(df)
}
\arguments{
\item{df}{dataframe with locations}
}
\value{
overwrites files in "C:/Work/Python" folders
}
\description{
A wrapper function for creating an ArcGIS map from location data
}
\details{
internal parameters of Python script specific to my BAEA GPS data.
  Identical to CreateArcGISandPDFMaps(), except it does not create PDFs.
}
