% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{CreateCategoricalLegend}
\alias{CreateCategoricalLegend}
\title{Creates a categorical legend for a map}
\usage{
Usage: CreateCategoricalLegend (metadata, metadata_layer, color_alpha,
  pos_x, pos_y, main, main_cex, main_col, lab_col, signif_digits, left,
  plot_new)
}
\arguments{
\item{metadata}{location of metadata .csv file. Metadata file must contain
"id" and "icon_color" columns.}

\item{metadata_layer}{not required, column name in metadata used for legend
labels}

\item{color_alpha}{two-digit rbga alpha value, "00" to "ff".}

\item{pos_x}{relative position of left and right edge of color bar on first
axis, [0,1]. Default is c(0.5, 0.55).}

\item{pos_y}{relative position on lower and upper edge of colar bar on second
axis, [0,1]. Default is c(0.05, 0.9).}

\item{main}{main title, written above the color bar. Default is NA.}

\item{main_cex}{relative size of main title. Default is 1.}

\item{main_col}{color of main title, default is "black".}

\item{lab_col}{color labels, default is "black".}

\item{left}{logical indicating whether to put the labels on the left (TRUE)
or on the right (FALSE). Default is FALSE.}

\item{plot_new}{logical indicating whether to create a new plot, default is
TRUE.}
}
\value{
Plot of legend for categorical data
}
\description{
Creates a legend for categorical data
}
\details{
Used in ExportKMLPolygon()
}
