% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{CreateCenterValueMatrix}
\alias{CreateCenterValueMatrix}
\title{Create a Matrix with a specific value in the center cell}
\usage{
CreateCenterValueMatrix(nrow, center_value)
}
\arguments{
\item{nrow}{integer, number of rows}

\item{center_value}{numeric, number in center cell}
}
\value{
a matrix
}
\description{
Create a Matrix with a specific value in the center cell
}
