% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{CreateColorIntervalSequence}
\alias{CreateColorIntervalSequence}
\title{Creates a color interval sequence for color ramps}
\usage{
CreateColorIntervalSequence(color_range, color_alpha, color_increment,
  log)
}
\arguments{
\item{color_range}{two-valued vector, the minimum and maximum values.}

\item{color_levels}{one value, number of color levels. Ignored if
color_interval not equal to NA.}

\item{color_increment}{one value, increment value of color breakpoints, one
value.}

\item{log}{logical of whether to log transform.}
}
\value{
vector of colors
}
\description{
Creates an interval sequence used for assigning color ramp values to objects.
}
\details{
Parameter defaults are object names because this function is used
  internally in CreateColorPaletteLegend() & ExportKMLPolygon() to ensure
  the interval sequence selection is identical. The code uses pretty(), so
  the returned color levels and legend levels may not be identical to the
  input parameter value.
}
