% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{CreateColorPaletteLegend}
\alias{CreateColorPaletteLegend}
\title{Creates a color palette legend for maps}
\usage{
CreateColorPaletteLegend (color_pal, color_range, color_alpha,
  color_levels, color_increment, legend_levels, log, pos_x, pos_y, main,
  main_cex, main_col, lab_col, signif_digits, left, plot_new)
}
\arguments{
\item{color_pal}{color palette to be used, also allowed are two extremes or
one value. Default is: c("yellow","red").}

\item{color_range}{two-valued vector, the minimum and maximum values.}

\item{color_alpha}{two-digit rbga alpha value, "00" to "ff", default is "ff".
Note: ExportKMLPOlygon() uses default "cc"}

\item{color_levels}{one value, number of color levels. Ignored if
color_interval not equal to NA. Default is 5.}

\item{color_increment}{one value, increment value of color breakpoints, one
value. Default is NA.}

\item{legend_levels}{one value, increment in legend values, ignored if
legend_values not equal to NA. Default is 5.}

\item{legend_values}{vector of specific legend labels. Default is NA.}

\item{log}{logical of whether to log transform. Default is FALSE.}

\item{pos_x}{relative position of left and right edge of color bar on first
axis, [0,1]. Default is c(0.5, 0.55).}

\item{pos_y}{relative position on lower and upper edge of colar bar on second
axis, [0,1]. Default is c(0.05, 0.9).}

\item{main}{main title, written above the color bar. Default is NA.}

\item{main_cex}{relative size of main title. Default is 1.}

\item{main_col}{color of main title. Default is "black".}

\item{lab_col}{color labels, default is "black".}

\item{signif_digits}{integer, number of signifcant digits}

\item{left}{logical indicating whether to put the labels on the left (TRUE)
or on the right (FALSE). Default is FALSE.}

\item{plot_new}{logical indicating whether to create a new plot. Default is
TRUE.}
}
\value{
Plot of legend with color palette
}
\description{
Creates a display of selected color palettes
}
\details{
Used in ExportKMLPolygon, make sure code used for color range,
  selction, log-tranformation, and color palette creation is identical or
  produces same results.  Color levels and legend levels use the pretty(), so
  the actual number of levels may not be identical to parameter values.
}
