% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kml.R
\name{CreateColorsByMetadata}
\alias{CreateColorsByMetadata}
\title{CreateColorsByMetadata}
\usage{
CreateColorsByMetadata(file = , metadata_id)
}
\arguments{
\item{file}{CSV file with columns for "id" and "icon_color"}

\item{metadata_id}{column name for unique identifier, default is "id"}
}
\value{
df with ID names and hexidecimal colors
}
\description{
Creates and/or displays dataframe of IDs and their associated colors
}
\details{
Used in several other functions
}
