% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{CreateExtentBuffer}
\alias{CreateExtentBuffer}
\title{Creates a buffer around a spatial extent}
\usage{
CreateExtentBuffer(df, buffer)
}
\arguments{
\item{df}{dataframe with columns: long_utm, lat_utm}

\item{buffer}{minimum buffer extent, default = 500}
}
\value{
vector of extent (long_min, long_max, lat_min, lat_max)
}
\description{
Create buffer around extent
}
\details{
finds lat and long extents, applies minimum buffer, and rounds to
  the same significance value as the buffer
}
