% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{CreateGaussKernRaster}
\alias{CreateGaussKernRaster}
\title{Create a Gaussian Kernel RasterLayer}
\usage{
CreateGaussKernRaster(sigma, nrow, shift_n, cell_size)
}
\arguments{
\item{sigma}{numeric, sigma (or sd) of distribution - scaled to cell_size}

\item{nrow}{numeric, number of rows/columns in RasterLayer}

\item{shift}{logical, whether to shift cells to center. If TRUE, center
middle cell at c(0,0). Default is FALSE.}

\item{shift_n}{numeric, shift in lat/long origin position}

\item{cell_size}{numeric, cell size of RasterLayer}
}
\value{
RasterLayer
}
\description{
Create a Gaussian Kernel RasterLayer
}
