% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{CreateKDEPoints}
\alias{CreateKDEPoints}
\title{Creates kernel density estimate (kde) raster based on spatial points
Creates a 'kde' object using a df of location points to calculate probability
  estimates for an input raster's cell centers}
\usage{
CreateKDERaster(df, df_lat, df_long, blank_raster)
}
\arguments{
\item{df}{dataframe of location data}

\item{df_long}{df col name of longitude values, default is "long_utm"}

\item{df_lat}{df col name of latitude values, default is "lat_utm"}

\item{buffer}{value used in CreateExtentBuffer() around df locations, default
is 500m}

\item{blank_raster}{file name or object name of raster file used to create
output. This raster is cropped based on the df's extent and kde estimates
are based on each cell's center. Default file is:
"C:/ArcGIS/Data/BlankRaster/maine_30mc.tif".}
}
\value{
A 'kde' object with a kde estimate for the center point of each blank
  raster cell that falls within the buffer around the df locations
}
\description{
Creates kernel density estimate (kde) raster based on spatial points
Creates a 'kde' object using a df of location points to calculate probability
  estimates for an input raster's cell centers
}
\details{
Ouput RasterLayer has the same CRS as the blank_raster input
}
