% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{CreateKDEProbs}
\alias{CreateKDEProbs}
\title{Creates a probability raster layer based on a 'kde' layer}
\usage{
ExportKMLProbs(kde_df, kde_raster, probs)
}
\arguments{
\item{kde_object}{'kde' object, usually made by CreateKDEPoints()}

\item{kde_raster}{'RasterLayer' created using estimates from a 'kde' object,
usually made by CreateKDERaster()}

\item{probs}{vector [0,1] of probability values, default value is seq(0, 1,
by=.1)}
}
\value{
A 'RasterLayer' with reclassified probabilities from using
  contourLevels() in the 'ks' package.
}
\description{
Calculates kernel probability levels based on 'kde' object and use them to
  recalssify the values in an kde 'RasterLayer' to the probability levels.
}
\details{
To be used between CreateKDERaster() and ExportKMLProbs()
}
