% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{CreateKDERaster}
\alias{CreateKDERaster}
\title{CreateKDERaster}
\usage{
CreateKDERaster(kde_points, cell_size, crs)
}
\arguments{
\item{kde_points}{dataframe of kde estimates at gridded points}

\item{cell_size}{cell size of output raster, default = 30}

\item{crs}{crs of output raster, default is UTM 19N, NAD83}
}
\value{
A 'RasterLayer' object with kde estimates for each of the blank
  raster cells that fall within the buffered df locations.
}
\description{
Creates a 'RasterLayer' based on a 'kde' object of gridded point location
  probability estimates
}
\details{
The kde_points file must have "eval.points" and "estimate", as
  created by CreateKDEPoints(). Returned 'RasterLayer' has the same CRS as
  blank_raster input.
}
