% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{CreateMapExtentBB}
\alias{CreateMapExtentBB}
\title{Creates an 'sf' polygon based an 'sf' object's extent and aspect ratio}
\usage{
CreateMapExtentBB(sf_object, ext, asp)
}
\arguments{
\item{sf_object}{'sf' object}

\item{ext}{extension factor of the extent, default = 1.15.}

\item{asp}{ratio of the extent (width/height), default = 1}
}
\value{
'sf' object
}
\description{
Creates an 'sf' polygon based an 'sf' object's extent and aspect ratio
}
\details{
Useful for making a main-map extent box within a overview map. Not
    guaranteed to be identical to a tmap's actual main map extent.
}
