% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{CreateMove}
\alias{CreateMove}
\title{Create 'move' object}
\usage{
CreateMove(df)
}
\arguments{
\item{df}{Dataframe with locations}
}
\value{
A 'move'object
}
\description{
A wrapper function for creating a 'move' object
}
\details{
Internal parameters set specifically for my BAEA GPS data.
}
