% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{CreateOSMBaseBB}
\alias{CreateOSMBaseBB}
\title{Create bounding box needed for downloading an 'OpenStreetMap' or 'rosm'
    baselayer}
\usage{
CreateOSMBaseBB(sf_obj, type = c("om_type"))
}
\arguments{
\item{sf_obj}{a 'sf' object, for determing bounding box coordinates}

\item{type}{character, either "om_type" or "rosm_type", default = "om_type"}
}
\value{
a vector (for type = "rosm_type") or list (for type = "om_type")
}
\description{
Create bounding box needed for downloading an 'OpenStreetMap' or 'rosm'
    baselayer
}
