% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{CreateProbIsoplethRaster}
\alias{CreateProbIsoplethRaster}
\title{CreateProbIsoplethRaster}
\usage{
CreateProbIsoplethRaster(df, buffer, cell_size, probs)
}
\arguments{
\item{df}{dataframe of location data}

\item{buffer}{value used in CreateExtentBuffer() around df locations, default
is 500 m}

\item{cell_size}{cell size of output raster, default = 30}

\item{probs}{vector [0,1] of probability values, default value is seq(0, 1,
by=.1)}
}
\value{
A Raster object with a probability estimate for each blank raster
  cell that falls within the buffer around the df locations
}
\description{
A wrapper function of used to create a RasterLayer of probabilities based on
  a df of location points
}
\details{
Ouput RasterLayer has the same CRS as the blank_raster input
}
