% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{CreateRasterConNestDistProb}
\alias{CreateRasterConNestDistProb}
\title{Creates a Raster of probability values based on Conspecific/Nest distance
    and adjusted based on the agent's current location}
\usage{
CreateRasterConNestDistProb(
  con_nest_raster,
  raster_extent,
  pars_gamma,
  pars_rescale,
  x,
  y,
  base
)
}
\arguments{
\item{con_nest_raster}{raster, conspecific and nest distance value raster}

\item{raster_extent}{extent, extent of the output raster}

\item{pars_gamma}{pars_gamma list (contains shape and rate)}

\item{pars_rescale}{pars_rescale list (contains y_min, y_max, y_min_new,
y_max_new}

\item{x}{integer x coordinate}

\item{y}{integer y coordinate}

\item{base}{raster, base}
}
\value{
Raster
}
\description{
Creates a Raster of probability values based on Conspecific/Nest distance
    and adjusted based on the agent's current location
}
