% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{CreateRasterFromPointsAndBase}
\alias{CreateRasterFromPointsAndBase}
\title{CreateRasterFromPointsAndBase}
\usage{
CreateRasterFromPointsAndBase(df, value, x, y, buffer, base)
}
\arguments{
\item{df}{input dataframe with columns for lat, long, and value}

\item{value}{column name with data for cell values}

\item{x}{column name for latitude coordinates}

\item{y}{column name for longitude coordinates}

\item{buffer}{distance in cell units to buffer around 'df' points for output
Rasterlayer. If NULL(default), the entire extent of the base parameter is
used for the output Rasterlayer.}

\item{base}{raster that is used to rasterize() the input data, default is set
to a RasterLayer specific to my BAEA project}
}
\value{
A 'RasterLayer' object with values in all the cells that had
  locations.
}
\description{
Creates a 'RasterLayer' from an input dataframe of point locations and a
  base RasterLayer that sets the CRS and cell locations for the output
}
\details{
Input latitude and longitude must match the CRS of base raster.
  Returned 'RasterLayer' has same CRS as base raster.
}
