% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{CreateRasterNestConDist}
\alias{CreateRasterNestConDist}
\title{CreateRasterNestConDist}
\usage{
CreateRasterNestConDist(years, buffer, study_nests, all_nests, base)
}
\arguments{
\item{years}{vector of years to calculate distances}

\item{buffer}{distance in cell units to buffer around nests for calculating
the distance to conspecifics. Default is 1000.}

\item{study_nests}{nests with "active_(year)", "eagle_id", "lat_utm", and
"long_utm" columns}

\item{all_nests}{nests with "active_(year)", "lat_utm", and "long_utm"}

\item{base}{raster that is used to rasterize() the input data, default is set
to a RasterLayer specific to my BAEA project}
}
\value{
A list of Rasters with nest and conspecific distances for all cells
  within buffer distance of the study nests. List is structured as:
  list[[year]][[nest_id]]
}
\description{
Creates a list of RasterLayers of distances to study nests (nests to create
  rasters for) and all nests (all conspecific nests in the area)
}
\details{
Some of the functions are hardwired with specific column names
}
