% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{CreateSpatialLines}
\alias{CreateSpatialLines}
\title{CreateSpatialLines}
\usage{
CreateSpatialLines(df, long, lat, crs)
}
\arguments{
\item{df}{dataframe with locations}

\item{long}{df column name of longitude, default is "long_utm"}

\item{lat}{df column name of latitude, default is "long_utm"}

\item{crs}{coordinate reference system, default is NA.}
}
\value{
a 'SpatialLines'object
}
\description{
A wrapper function for creating a 'SpatialLines' object
}
