% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{CreateSpatialPolygons}
\alias{CreateSpatialPolygons}
\title{CreateSpatialPolygons}
\usage{
CreateSpatialPolygons(df, long, lat, crs)
}
\arguments{
\item{df}{dataframe with locations}

\item{long}{df column name of longitude, default is "long_utm"}

\item{lat}{df column name of latitude, default is "long_utm"}

\item{crs}{coordinate reference system, default is NA.}
}
\value{
a 'SpatialPolygons'object
}
\description{
A wrapper function for creating a 'SpatialPolygons' object
}
