% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{download_shapefiles}
\alias{download_shapefiles}
\title{Download shapefile zipfile from googledrive}
\usage{
download_shapefiles(
  country,
  org_label = "country",
  drive_folder = NULL,
  dest_file = NULL,
  overwrite = TRUE,
  unzip = FALSE
)
}
\arguments{
\item{country}{PEPFAR Countryname}

\item{org_label}{Orgunit label, default is set to country}

\item{drive_folder}{Googledrive id for all PEPFAR Spatial files}

\item{dest_file}{Full file name where to download zipped shapefile}

\item{overwrite}{Should the process overwrite existing files}

\item{unzip}{Should the zipfile be unzipped}
}
\description{
Download shapefile zipfile from googledrive
}
\examples{
\dontrun{

  cntry <- "Zambia"

  download_shapefiles(country = cntry)

  download_shapefiles(country = cntry, org_label = "snu1", unzip = TRUE)

  download_shapefiles(country = cntry, org_label = "psnu", unzip = TRUE)
}

}
