% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{dview}
\alias{dview}
\title{View feature attributes}
\usage{
dview(geodata, console = FALSE)
}
\arguments{
\item{geodata}{spatial data data}

\item{console}{view in console? default false}
}
\description{
View feature attributes
}
\examples{
\dontrun{
gview(sf_data)
}

}
