% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_locations.R
\name{explore_facilities}
\alias{explore_facilities}
\title{Explore facility locations}
\usage{
explore_facilities(.data, cntry, terr_path = NULL)
}
\arguments{
\item{.data}{Datim organisation units data frame}

\item{cntry}{Country name}

\item{terr_path}{Path to terrain raster dataset}
}
\description{
Explore facility locations
}
\examples{
\dontrun{
explore_facilities(df_sites, cntry = "saturn")
df_sites \%>\% explore_facilities("saturn")
}

}
