% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kml.R
\name{ExportKMLPolygon}
\alias{ExportKMLPolygon}
\title{ExportKMLPolygon}
\usage{
ExportKMLPolygon(object, object_layer, outfile, kml_name,
  categorical, metadata, metadata_layer, color_pal, color_alpha, color_range,
  color_min, color_max, color_levels, color_increment, legend_levels,
  legend_values, log, signif_digtits, outline, alt_mode, extrude, labelscale,
  create_kmz, ...)
}
\arguments{
\item{object}{a 'SpatialPolygonsDataFrame' object}

\item{object_layer}{layer in object that is used for display}

\item{outfile}{location of output KML file. Extensions (.kml or .kmz) will
automatically determine file type.}

\item{kml_name}{name of folder in KML file}

\item{categorical}{logical, whether the data is categorical, default is
false.}

\item{metadata}{location of metadata .csv file for categorical data.
Metadata file must contain "id" and "icon_color" columns.}

\item{metadata_layer}{not required, column name in metadata used for legend
labels instead of "id"}

\item{color_pal}{color palette, can be a color ramp (i.e. c("white", "red")
or a specific palette ("SAGA_pal[[1]]")}

\item{color_alpha}{display and legend alpha value, default "cc"}

\item{color_range}{range of object values to create color palette}

\item{color_min}{min object values to create color palette}

\item{color_max}{max object value to create color palette}

\item{color_levels}{number of breaks in color palette, ignored if
color_interval not equal to NA}

\item{color_increment}{intervals for color palette breaks}

\item{legend_levels}{number of breakpoints in legend, ignored if
legend_interval not equal to NA}

\item{legend_values}{vector of values in legend}

\item{log}{logical of whether to log transform. Default is FALSE.}

\item{signif_digits}{number of signifcant digits for polygon labels}

\item{outline}{1 or 0, whether to draw an outline around each polygon}

\item{alt_mode}{based on KML code: "absolute","clampedToGround",
"relativeToGround" (see KML documentation for description). Default is
"clampedToGround"}

\item{extrude}{either 0 (default) or 1: 0 if for no line, 1 extends a line
from the point to the ground.}

\item{labelscale}{adjusts the size of the Google Earth location point
labels. Default is 0, which hides the labels. To show labels, change to a
value between 0.7-1.}

\item{create_kmz}{will always create a KMZ, default is FALSE.}

\item{...}{parameters passed to CreateCreateCategoricalLegend() or
CreateColorPaletteLegend(), e.g., legend_levels, legend_values}
}
\value{
KML or KMZ of polygons with an associated legend
}
\description{
Create a Google Earth KML file from a SpatialPolygonsDataFrame
}
\details{
uses CreateCategoricalLegend() or CreateColorPaletteLegend()
  depending on if categorical = TRUE or FALSE
}
