% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kml.R
\name{ExportKMLProbContour}
\alias{ExportKMLProbContour}
\title{ExportKMLProbContour}
\usage{
ExportKMLProbContour(kde_raster, kml_name, dissolve, ...)
}
\arguments{
\item{probs_raster}{RasterLayer of kde probability levels, usually made with
CreateHRProbContourRaster()}

\item{kml_name}{name of folder in KML file, default= "prob_countours"}

\item{dissolve}{logical, TRUE = dissolve adjacent like polygons,
FALSE = each cell remains intact; default is FALSE}

\item{...}{other arguments to pass to ExportKMLPolygon()}
}
\value{
Exports a KML or KMZ of the probability contour levels
}
\description{
Create a Google Earth KML or KMZ file based on a 'RasterLayer' of kernel
  probability levels
}
\details{
Use the dissolve functionality with caution - it is unable to
  properly handle polygon holes. This results from a known problem regarding
  holes in 'Polygon' objects. Therefore, at a given contour level holes may
  be improperly filled in.
}
