% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kml.R
\name{ExportKMLRasterOverlayWithTime}
\alias{ExportKMLRasterOverlayWithTime}
\title{ExportKMLRasterOverlayWithTime}
\usage{
ExportKMLRasterOverlayWithTime(raster, time, color_pal, alpha,
  maxpixels, blur, colNA, outfile, output_dir)
}
\arguments{
\item{raster}{a Raster* object}

\item{time}{Interval* object, default = NULL}

\item{color_pal}{color palette, can be a color ramp (e.g., c("white", "red")
or a specific palette (e.g., "SAGA_pal[[1]]")}

\item{alpha}{numeric (0-1), transparency level of the kml. Default is 1.}

\item{method}{method used to compute values for the new RasterLayer.
Either 'ngb' (nearest neighbor), which is useful for categorical
variables, or 'bilinear' (bilinear interpolation; the default value),
which is appropriate for continuous variables.}

\item{overwrite}{logical, whether or not to overwrite result file. Default
is TRUE.}

\item{maxpixels}{maximum number of pixels. If ncell(raster) > maxpixels,
sample is used to reduce the number of pixels.}

\item{blur}{integer (default=10). Higher values help avoid blurring of
isolated pixels (at the expense of a png file that is blur^2 times larger)}

\item{colNA}{color to use for the background (default is transparent)}

\item{outfile}{name of KML, default is to use name of raster}

\item{output_dir}{output folder location, default is getwd()}

\item{zip}{logical, whether or not to convert .kml to .kmz}
}
\value{
KML of a Raster
}
\description{
Exports kml of a raster with time properties
}
\details{
Modified from functions in the 'kml' and 'raster' packages
}
