% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kml.R
\name{ExportKMLTelemetry}
\alias{ExportKMLTelemetry}
\title{ExportKMLTelemetry}
\usage{
ExportKMLTelemetry(df, id, datetime, lat, long, alt, alt_mode, speed,
  agl, behavior, point_color, point_metadata, point_pal, point_r_pal,
  point_b_pal, extrude, path, path_color, path_metadata, path_pal,
  path_r_pal, path_b_pal, arrow, icon_by_sex, labelscale, dateformat,
  timeformat, datetimeformat, file, output_dir)
}
\arguments{
\item{df}{input dataframe, must have id, lat, long, and datetime}

\item{id}{column name of unique identifier, data is split into unique paths
and separate folders based on this parameter}

\item{datetime}{column name of datetime in POSIXct format or as a character
in the format (\%Y/\%m/\%d \%H:\%M)}

\item{lat}{column name of latitude coordinates (WGS84, dec. degree)}

\item{long}{column name of longitude coordinates (WGS84, dec. degree)}

\item{alt}{input dataframe column name for altitude(m). Optional.}

\item{alt_mode}{based on KML code: "absolute","clampedToGround",
"relativeToGround" (see KML documentation for description). Default is
"clampedToGround".}

\item{speed}{input dataframe column name for speed. Optional}

\item{agl}{input dataframe column name for "altitude above ground level",
optional}

\item{behavior}{input dataframe column name for behavior. Optional}

\item{point_color}{column name that determines the color for each point, may
be same as 'id' parameter, but may also be sex, behavior, season, etc.
Default is 'id' parameter}

\item{point_metadata}{location of metadata .csv file. Metadata file must
have a column that matches name of 'point_color'parameter and "icon_color"
column with hexadecimal colors.}

\item{point_pal}{name of color palette funtions (e.g., rainbow, heat.colors,
terrain.colors, topo.colors, cm.colors used to create colors. This
parameter has priority over the other point color palette parameters.
Default is NULL.}

\item{point_r_pal}{Specifc number of 'R_pal' color palette from the
'PlotKML' Package (e.g., 1 = R_pal[[1]]). This parameter has priority over
the 'b_pal' parameter for setting the colors. Default is NULL.}

\item{point_b_pal}{color palette name from RColorBrewer package, default is
"Set1". Automatically adjusts number of colors to match the unique number
of factors in the 'point_color' column of the input dataframe.}

\item{extrude}{logical, either FALSE (default) for no line, or TRUE which
extends a line from the point to the ground.}

\item{path}{logical, to create Track paths. Default is TRUE.}

\item{path_color}{similar to 'point_color' parameter, but the value must
have the same factor level structure as the id file, because each path is
constructed for each id factor. Default will use 'id' parameter.}

\item{path_metadata}{location of metadata .csv file. Metadata file must have
a column that matches name of 'path_color' parameter and an "icon_color"
column with hexadecimal colors.}

\item{path_pal}{name of color palette funtions (e.g., rainbow, heat.colors,
terrain.colors, topo.colors, cm.colors used to create colors. This
parameter has priority over the other point color palette parameters.
Default is NULL.}

\item{path_r_pal}{Specifc number of 'R_pal' color palette from the 'PlotKML'
Package (e.g., 1 = R_pal[[1]]). This parameter has priority over the
'b_pal' parameter for setting the colors. Default is NULL.}

\item{path_b_pal}{color palette name from RColorBrewer package, default is
"Set1". Automatically adjusts number of colors to match the unique number
of factors in the 'point_color' column of the input dataframe.}

\item{arrow}{logical, use arrow heads on path icons.}

\item{icon_by_sex}{logical, use different icons based on "sex" column}

\item{labelscale}{adjusts the size of the Google Earth location point
labels. Default is 0, which hides the labels. To show labels, change to a
value between 0.7-1.}

\item{dateformat}{changes the format of the date in the Google Earth
location pop-up windows. Default is "\%Y/\%m/\%d".}

\item{timeformat}{changes the format of the time in the Google Earth
locations pop-up windows. Default is "\%I:\%M \%p".}

\item{datetimeformat}{changes the datetime format for the label of
highlighted points. Default is "\%Y/\%m/\%d \%I:\%M \%p"}

\item{file}{filename of output KML file, default is name of input dataframe}

\item{output_dir}{name for folder in the KML file, default is working
directory}
}
\value{
KML of points and multitracks
}
\description{
Create a Google Earth KML file (points and multitrack) from lat/long
  coordinates.
}
