% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kml.R
\name{ExportKMLTelemetryBAEA}
\alias{ExportKMLTelemetryBAEA}
\title{Export BAEA telemetry data at KML}
\usage{
ExportKMLTelemetryBAEA(df, id, datetime, lat, long, alt, alt_mode,
  speed, agl, behavior, point_color, point_metadata, point_pal, point_r_pal,
  point_b_pal, extrude, path, path_color, path_metadata, path_pal,
  path_r_pal, path_b_pal, arrow, icon_by_sex, labelscale, dateformat,
  timeformat, datetimeformat, file, output_dir)
}
\arguments{
\item{df}{Input dataframe, must have id, lat, long, and datetime.}

\item{id}{Column name of unique identifier, data is split into unique paths
and separate folders based on this parameter. Default is "id".}

\item{datetime}{Column name of datetime in POSIXct format or as a character
in the format (\%Y/\%m/\%d \%H:\%M). Default is "datetime".}

\item{lat}{Column name of latitude coordinates (WGS84, dec. degree). Default
"lat".}

\item{long}{Column name of longitude coordinates (WGS84, dec. degree).
Default is "long".}

\item{alt}{Input dataframe column name for altitude(m). Optional. Default is
"alt".}

\item{alt_mode}{based on KML code: "absolute","clampedToGround",
"relativeToGround" (see KML documentation for description). Default is
"absolute".}

\item{speed}{input dataframe column name for speed. Optional. Default is
"speed".}

\item{agl}{input dataframe column name for "altitude above ground level",
optional. Default is NULL.}

\item{behavior}{input dataframe column name for behavior. Optional. Default
is NULL.}

\item{point_color}{column name that determines the color for each point, may
be same as 'id' parameter, but may also be sex, behavior, season, etc.
Default is 'deploy_location".}

\item{point_metadata}{location of metadata .csv file. Metadata file must
have a column that matches name of 'point_color'parameter and "icon_color"
column with hexadecimal colors. Default is:
"Data/GPS/GPS_Deployments.csv".}

\item{point_pal}{name of color palette funtions (e.g., rainbow, heat.colors,
terrain.colors, topo.colors, cm.colors used to create colors. This
parameter has priority over the other point color palette parameters.
Default is NULL.}

\item{point_r_pal}{Specifc number of 'R_pal' color palette from the
'PlotKML' Package (e.g., 1 = R_pal[[1]]). This parameter has priority over
the 'b_pal' parameter for setting the colors. Default is NULL.}

\item{point_b_pal}{color palette name from RColorBrewer package, default is
"Set1". Automatically adjusts number of colors to match the unique number
of factors in the 'point_color' column of the input dataframe.}

\item{extrude}{Logical, either FALSE (default) for no line, or TRUE which
extends a line from the point to the ground.}

\item{path}{Logical, to create Track paths. Default is TRUE.}

\item{path_color}{Similar to 'point_color' parameter, but the value must
have the same factor level structure as the id file, because each path is
constructed for each id factor. Default is "deploy_location".}

\item{path_metadata}{Location of metadata .csv file. Metadata file must have
a column that matches name of 'path_color' parameter and an "icon_color"
column with hexadecimal colors. Default is:
"Data/GPS/GPS_Deployments.csv".}

\item{path_pal}{Name of color palette funtions (e.g., rainbow, heat.colors,
terrain.colors, topo.colors, cm.colors used to create colors. This
parameter has priority over the other point color palette parameters.
Default is NULL.}

\item{path_r_pal}{Specifc number of 'R_pal' color palette from the 'PlotKML'
Package (e.g., 1 = R_pal[[1]]). This parameter has priority over the
'b_pal' parameter for setting the colors. Default is NULL.}

\item{path_b_pal}{color palette name from RColorBrewer package, default is
"Set1". Automatically adjusts number of colors to match the unique number
of factors in the 'point_color' column of the input dataframe.}

\item{arrow}{Logical, use arrow heads on path icons. Default is TRUE.}

\item{icon_by_sex}{Logical, use different icons based on "sex" column.
Default is FALSE.}

\item{labelscale}{Numeric, adjusts the size of the Google Earth location
point labels. Default is 0, which hides the labels. To show labels, change
to a value between 0.7-1.}

\item{dateformat}{String, changes the format of the date in the Google Earth
location pop-up windows. Default is "\%Y/\%m/\%d".}

\item{timeformat}{String, changes the format of the time in the Google Earth
locations pop-up windows. Default is "\%I:\%M \%p".}

\item{datetimeformat}{String, changes the datetime format for the label of
highlighted points. Default is "\%Y/\%m/\%d \%I:\%M \%p"}

\item{file}{String, filename of output KML file, default is name of input
dataframe. Default is: "BAEA Data.kml"}

\item{output_dir}{Name for folder in the KML file, default is working
directory. Default is: "C:/Users/Blake/Desktop".}
}
\value{
KML of points and multitracks
}
\description{
Create a Google Earth KML file (points and multitrack) from lat/long
  coordinates of BAEA telemetry data. Uses the same functionality as
  ExportKMLTelemetry(), but all the defaults are set for specific BAEA
  telemetry data from Maine
}
