% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kml.R
\name{ExportKMLWindTurbines}
\alias{ExportKMLWindTurbines}
\title{ExportKMLWindTurbines}
\usage{
ExportKMLWindTurbines(df, labelscale, outfile, metadata)
}
\arguments{
\item{df}{the location of the .csv file}

\item{labelscale}{adjusts the size of the Google Earth location point
labels. Default is 0, which hides the labels. To show labels, change to a
value between 0.7-1.}

\item{outfile}{location of output KML file}

\item{metadata}{ocation of metadata .csv file. Metadata file must contain
columns for project ID, hexadecimal colors, and additional icon data.}
}
\value{
creates KML
}
\description{
Exports KML of wind turbine locations
}
\details{
Defaults are specific to my file directories and locations
}
