% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{ExportRasterNestConDist}
\alias{ExportRasterNestConDist}
\title{ExportRasterNestConDist}
\usage{
ExportRasterNestConDist(nest_con_dist, dir)
}
\arguments{
\item{nest_con_dist}{list of RasterLayers of distances with the structure of
raster[[year]][[nest_id]]}

\item{dir}{directory for the output raster files}
}
\value{
Exports rasters to directory location
}
\description{
Creates a list of RasterLayers of distances to study nests (nests to create
  rasters for) and all nests (all conspecific nests in the area)
}
\details{
Some of the functions are hardwired with specific column names
}
