% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{ExportShapefileFromPoints}
\alias{ExportShapefileFromPoints}
\title{ExportShapefileFromPoints}
\usage{
ExportShapefileFromPoints(df, lat, long, name, folder, crs, overwrite)
}
\arguments{
\item{df}{dataframe with locations}

\item{lat}{latitude column name, in same coordinates as 'crs'. Default
"lat"}

\item{long}{longitude column name, in same coordinates as 'crs'. Default
"long"}

\item{name}{name of shapefile output}

\item{folder}{location for shapefile files}

\item{crs}{coordinate reference system, default is
"+proj=longlat +datum=WGS84"}

\item{overwrite}{logical, overwrite outfile. Use with caution, default FALSE.}
}
\value{
creates shapefile
}
\description{
Exports shapefile of a dataframe's location data
}
\details{
requires "lat" and "long" columns
}
