% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{extract_boundaries}
\alias{extract_boundaries}
\title{Extract PEPFAR Orgunit Boundaries}
\usage{
extract_boundaries(
  spdf,
  country,
  level = 3,
  username,
  password,
  export = FALSE,
  name = NULL
)
}
\arguments{
\item{spdf}{PEPFAR Global Shapefile}

\item{country}{Country name}

\item{level}{Orgunit level}

\item{username}{Datim username}

\item{password}{Datim password}

\item{export}{Export extract as shapefile?}

\item{name}{Export filename}
}
\value{
sf object with orgunit attributes
}
\description{
PEPFAR VcPolygons are shared with orgunit uids only,
making hard for analysts to identify specific polygon
each orgunit level. This function extract orgunt attributes
from Datim and append them to the global shapefile, allowing
analysts to filter and work only with subset.
}
\examples{
\dontrun{

 library(gisr)
 library(sf)
 library(glamr)

 cntry <- "Tanzania"

 shp_pepfar <- get_vcpolygons(folderpath = glamr::si_path("path_vector")) \%>\%

 shp_country <- extract_boundaries(
   spdf = shp_pepfar,
   country = cntry,
   level = 3
 )

 shp_country \%>\% gview()}


}
