% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_locations.R
\name{extract_facilities}
\alias{extract_facilities}
\title{Extract facility sites}
\usage{
extract_facilities(.data)
}
\arguments{
\item{.data}{Datim organisation units data frame - with label and coordinates columns}
}
\description{
Extract facility sites
}
\examples{
\dontrun{
  grabr::datim_pull_hierarchy(...) \%>\%
    extract_facilities()
}

}
