% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_locations.R
\name{extract_locations}
\alias{extract_locations}
\title{Extract location data}
\usage{
extract_locations(country, username, password)
}
\arguments{
\item{country}{PEPFAR Operating Unit or Regional Countries}

\item{username}{Datim Account Username}

\item{password}{Datim Account Key}
}
\description{
Extract location data
}
\examples{
\dontrun{
extract_locations("saturn", "my_username", "my_password")
}
}
