% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_map.R
\name{extract_raster}
\alias{extract_raster}
\title{Extract raster data for an AOI (Countries)}
\usage{
extract_raster(countries, mask = FALSE, buffer = 0.1, ras = NULL)
}
\arguments{
\item{countries}{List of the country names or sf object}

\item{mask}{Should the extracted data match the exact boundary limits?}

\item{buffer}{Extend AOI extent by x}

\item{ras}{RasterLayer or Path to raster file}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      spdf spatial dataframe
}\if{html}{\out{</div>}}
}
\description{
Extract raster data for an AOI (Countries)
}
\examples{
\dontrun{
 library(gisr)
 library(sf)

 get_terrain(countries = list("Zambia"))
 get_terrain(countries = list("Zambia"), mask = TRUE)
 get_terrain(countries = list("Zambia"), buffer = .5, terr = "../../HDX_Data")
}

}
