% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{extract_roads}
\alias{extract_roads}
\title{Extract Road Network data from OSM}
\usage{
extract_roads(aoi, radius = NULL, clip = FALSE)
}
\arguments{
\item{aoi}{Area of Interest as sf object}

\item{radius}{Buffer redius in meters, default = 1000m}

\item{clip}{Should the output be clipped to the AOI? Default is FALSE}
}
\description{
Extract Road Network data from OSM
}
\examples{
\donttest{

 library(gisr)

 adm_zmb <- get_admin1(countries = "Zambia") \%>\%
   dplyr::select(name) \%>\%
   dplyr::filter(name == 'Lusaka')

 adm_zmb \%>\%
   extract_roads() \%>\%
   gview()
}

}
