% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{ExtractMatrixCenterRow}
\alias{ExtractMatrixCenterRow}
\title{Extract values in center row of matrix and convert to a dataframe}
\usage{
ExtractMatrixCenterRow(matrix_in)
}
\arguments{
\item{matrix_in}{matrix}
}
\value{
matrix
}
\description{
Extract values in center row of matrix and convert to a dataframe
}
