% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{ExtractRasterCenterRow}
\alias{ExtractRasterCenterRow}
\title{Extract values in center row of RasterLayer and convert to a dataframe}
\usage{
ExtractRasterCenterRow(raster_in)
}
\arguments{
\item{raster_in}{matrix}
}
\value{
matrix
}
\description{
Extract values in center row of RasterLayer and convert to a dataframe
}
