% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{gattributes}
\alias{gattributes}
\title{Get attributes from feature class}
\usage{
gattributes(geodata)
}
\arguments{
\item{geodata}{spatial data}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    attribures as data frame
}\if{html}{\out{</div>}}
}
\description{
Get attributes from feature class
}
\examples{
\dontrun{
 library(gisr)

 # Admin level 0 [country] geodata
 adm0 <- get_admin0("Ghana")

 # Extract attrs from geodata
 .df <- gattributes(geodata = adm0)

 head(.df)
}

}
