% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_reports.R
\name{generate_facilities_report}
\alias{generate_facilities_report}
\title{Report locations data completeness}
\usage{
generate_facilities_report(
  cntry,
  mer_sites,
  user,
  pass,
  terr_path = NULL,
  output_folder = NULL
)
}
\arguments{
\item{cntry}{Country name}

\item{mer_sites}{Latest MER Sites Results / Targets (cols: orgunituid, sitename)}

\item{user}{Datim account username}

\item{pass}{Datim account password (glamr::mypwd is recommended)}

\item{terr_path}{Path to terrain raster data}

\item{output_folder}{Output folder}
}
\description{
Report locations data completeness
}
\examples{
\dontrun{
generate_sites_report(cntry = "saturn", mer_sites = sites)
}

}
