% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hexbins.R
\name{generate_hexbins}
\alias{generate_hexbins}
\title{Generate an hexbins polygon sfc from another sfc}
\usage{
generate_hexbins(geodata, size = NULL)
}
\arguments{
\item{geodata}{r spatial data}

\item{size}{size of each hex bin in meters}
}
\value{
country hex polygon as feature class
}
\description{
Generate an hexbins polygon sfc from another sfc
}
\examples{
\dontrun{
 library(gisr)

 cntry <- "Nigeria"
 data <- get_admin0(cntry)

 generate_hexbins(data)
 generate_hexbins(geodata = data, size = 10000)
}

}
