% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{geo_fence}
\alias{geo_fence}
\title{Generate a buffer around an Area of Interest}
\usage{
geo_fence(aoi, radius = 1000, append = TRUE)
}
\arguments{
\item{aoi}{Area of Interest as sf object}

\item{radius}{Buffer redius in meters, default = 1000m}

\item{append}{Should the buffered area be appended to the AOI? Default is TRUE}
}
\value{
simple feature class
}
\description{
Generate a buffer around an Area of Interest
}
\examples{
\dontrun{
 library(gisr)

 adm <- get_admin0(countries = "Zambia")

 adm \%>\% geo_fence(radius = 5000, append = TRUE) \%>\% gview()

 adm \%>\% geo_fence(radius = 5000, append = FALSE) \%>\% gview()
}

}
