% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{geo_neighbors}
\alias{geo_neighbors}
\title{Get neighbors of a given contry}
\usage{
geo_neighbors(src, countries, var = "sovereignt", crs = 4326, crop = FALSE)
}
\arguments{
\item{src}{Source spatial data frame}

\item{countries}{countries of interest}

\item{var}{Variable name}

\item{crs}{Coordinates reference system, default is WGS84 (EPGS:4326)}

\item{crop}{Crop sfc to focus countries extent?}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      simple feature class
}\if{html}{\out{</div>}}
}
\description{
Get neighbors of a given contry
}
\examples{
\dontrun{
 library(gisr)

 cntry <- "Zambia"

 shp_ne <- get_necountries()

 # Country + neighbors boundaries
 cntries <- geo_neighbors(src = shp_ne, countries = cntry)

 # Country + neighbors boundaries: crop to country extent
 cntries <- geo_neighbors(src = shp_ne, countries = cntry, crop = TRUE)

 cntries \%>\% gview()
}

}
