% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{geo_viz}
\alias{geo_viz}
\title{Plot polygon features}
\usage{
geo_viz(geodata)
}
\arguments{
\item{geodata}{r spatial data}
}
\description{
Plot polygon features
}
\examples{
\dontrun{
geo_viz(sf_data)
}

}
