% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adm_boundaries.R
\name{get_adm_boundaries}
\alias{get_adm_boundaries}
\title{Get geospatial data from GADM}
\usage{
get_adm_boundaries(
  country_code,
  adm_level = 0,
  download = TRUE,
  folderpath = NULL
)
}
\arguments{
\item{country_code}{ISO3 code}

\item{adm_level}{country administrative boundaries level}

\item{download}{Save a copy}

\item{folderpath}{where to save / read data from}
}
\value{
geodata as sf with crs = 4326
}
\description{
This function extracts / reads geospatial data
for specified country and administrative boundary levels
}
\examples{
\dontrun{
library(gisr)
get_adm_boundaries("UGA")
get_adm_boundaries("UGA", adm_level = 2, folderpath = "GIS")
}

}
