% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{get_admin1}
\alias{get_admin1}
\title{Get admin level 1 boundaries sf data for a given country}
\usage{
get_admin1(countries, crs = 4326)
}
\arguments{
\item{countries}{list of country names}

\item{crs}{coordinates reference system}
}
\value{
simple feature class
}
\description{
Get admin level 1 boundaries sf data for a given country
}
\examples{
\dontrun{
get_admin1(countries = list("Zambia"))
}

}
