% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{get_nepolygons}
\alias{get_nepolygons}
\title{Get Natural Earth Polygons}
\usage{
get_nepolygons(
  scale = c("large", "small", "medium"),
  type = c("countries", "map_units", "sovereignty", "tiny_countries")
)
}
\arguments{
\item{scale}{Scale of the map - options are 'large', 'small', 'medium'}

\item{type}{country type - options are 'countries', 'map_units', 'sovereignty', 'tiny_countries'}
}
\value{
world countries as sf object
}
\description{
Get Natural Earth Polygons
}
\examples{
\dontrun{
 library(gisr)

 shp_ne <- get_nepolygons()
 shp_ne <- get_nepolygons(scale = "large", type = "countries")
}

}
