% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_ouleveluids}
\alias{get_ouleveluids}
\title{Get Org level uids}
\usage{
get_ouleveluids(ouuid, level, username = NULL, password = NULL)
}
\arguments{
\item{ouuid}{Operatingunit uid}

\item{level}{Orgunit level}

\item{username}{Datim Account username}

\item{password}{Datim Account Password}
}
\value{
list of uids
}
\description{
Get Org level uids
}
\examples{
\dontrun{
get_ouleveluids("<ou-uid-goes-here>", level = 4)
}

}
