% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_ouorgs}
\alias{get_ouorgs}
\title{Get list of OU Orgs at specific level}
\usage{
get_ouorgs(
  ouuid,
  level = 4,
  username = NULL,
  password = NULL,
  baseurl = "https://final.datim.org/"
)
}
\arguments{
\item{ouuid}{OU uid}

\item{level}{org level}

\item{username}{DATIM Username}

\item{password}{DATIM password, recommend using \code{mypwd()}}

\item{baseurl}{base url for the API, default = https://final.datim.org/}
}
\value{
ORG UIDS as tibble
}
\description{
Get list of OU Orgs at specific level
}
\note{
Use \code{get_orguids()} for levels above 4
}
\examples{
\dontrun{
 orgs <- get_ouorgs(ouuid = "<ou-uid-goes-here>", level = 4)}

}
