% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_ouorguids}
\alias{get_ouorguids}
\title{Get Orgs uids by level}
\usage{
get_ouorguids(ouuid, level, username = NULL, password = NULL)
}
\arguments{
\item{ouuid}{Operatingunit uid}

\item{level}{Orgunit level}

\item{username}{Datim Account username}

\item{password}{Datim Account Password}
}
\value{
list of uids
}
\description{
Get Orgs uids by level
}
\examples{
\dontrun{
get_ouorgsuids("<ou-uid-goes-here>", level = 4)
}

}
