% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_ouuid}
\alias{get_ouuid}
\title{Get OU Identifiers}
\usage{
get_ouuid(country, username, password)
}
\arguments{
\item{country}{Country name}

\item{username}{Datim Account username}

\item{label}{Keychain item name}
}
\description{
Get OU Identifiers
}
\examples{
\dontrun{
get_ouuid("Kenya", "my_username")
get_ouuid("Kenya", username = "my_username", label = "my_keychain_label")
}

}
