% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_ouuids}
\alias{get_ouuids}
\title{Get OU UIDS}
\usage{
get_ouuids(
  username = NULL,
  password = NULL,
  baseurl = "https://final.datim.org/"
)
}
\arguments{
\item{username}{DATIM Username, recommend using `datim_user()`}

\item{password}{DATIM password, recommend using `datim_pwd()`}

\item{baseurl}{base url for the API, default = https://final.datim.org/}
}
\value{
OU UIDS as tibble
}
\description{
Get OU UIDS
}
\examples{
\dontrun{
 ous <- get_ouuids() }

}
