% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{GetIntersectionArea}
\alias{GetIntersectionArea}
\title{Get Intersection Area
Gets the overlapping area of two sf polygon objects}
\usage{
GetIntersectionArea(poly_a, poly_b)
}
\arguments{
\item{poly_a}{sf polygon object}

\item{poly_b}{sf polygon object}
}
\value{

}
\description{
Get Intersection Area
Gets the overlapping area of two sf polygon objects
}
\details{
When two polygons have no overlap, the result is 0 in the units of
    the respective polygons
}
