% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{gview}
\alias{gview}
\title{Plot polygon features}
\usage{
gview(geodata)
}
\arguments{
\item{geodata}{r spatial data}
}
\description{
Plot polygon features
}
\examples{
\dontrun{
gview(sf_data)
}

}
