% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{ImportBAEA}
\alias{ImportBAEA}
\title{Imports 'baea' data}
\usage{
ImportBAEA(existing, import, tz)
}
\arguments{
\item{existing}{Dataframe, exisiting file to merge with baea. Can be NULL.
Default is "deployed".}

\item{import}{Logical, whether or not to import baea.csv file. Default is
TRUE.}

\item{tz}{String, timezone. Default is "Etc/GMT+5".}
}
\value{
Merged baea dataframe is returned. Also, writes new "baea.csv" file.
}
\description{
Imports baea.csv and merges with existing data
}
\details{
Directory defaults are specific to my computer
}
