% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{ImportLandscapeRasterStack}
\alias{ImportLandscapeRasterStack}
\title{ImportLandscapeRasterStack}
\usage{
ImportLandscapeRasterStack()
}
\value{
The original locations df with additional landscape data columns.
}
\description{
Imports raster layers and combines them into a RasterStack
}
\details{
This function is specific to my directory and datafiles
}
