% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{ImportUnits}
\alias{ImportUnits}
\title{Import units}
\usage{
ImportUnits(units = "deployed", existing = NULL, import = TRUE)
}
\arguments{
\item{units}{String, units to download either "deployed" or "reserve".}

\item{existing}{Exisiting file to merge with import records.}

\item{import}{Logical, whether to import previous records. Default is TRUE,}
}
\value{
Dataframe
}
\description{
Imports location records and merges with existing
}
\details{
Defaults are specific to my file directories and locations
}
