% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{KeepMatrixCenterRow}
\alias{KeepMatrixCenterRow}
\title{Keep values only in center row of matrix (other cells converted to zero)}
\usage{
KeepMatrixCenterRow(matrix_in)
}
\arguments{
\item{matrix_in}{matrix}
}
\value{
matrix
}
\description{
Keep values only in center row of matrix (other cells converted to zero)
}
