% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{KeepRasterCenterRow}
\alias{KeepRasterCenterRow}
\title{Keep values only in center row of Raster (convert other cells to zero)}
\usage{
KeepRasterCenterRow(raster_in)
}
\arguments{
\item{raster_in}{Raster}
}
\value{
Raster
}
\description{
Keep values only in center row of Raster (convert other cells to zero)
}
