% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{PackCircles}
\alias{PackCircles}
\title{PackCircles}
\usage{
PackCircles(config, extent, edge_buffered, max_iter, overlap)
}
\arguments{
\item{config}{matrix with two cols: radius, n}

\item{extent}{extent of bounding rectangle: c(x_min, x_max, y_min,y_max).
Default is (0, 10000, 0, 10000).}

\item{edge_buffer}{distance from edge of extent that circle center's must be
placed}

\item{max_iter}{maximum number of iterations to try}

\item{overlap}{allowable overlap expressed as proportion of joint radii}

\item{plot}{logical, whether or not to draw a plot}
}
\value{
A dataframe of points
}
\description{
Simple circle packing algorithm based on inverse size weighted repulsion
}
\details{
Original code from: http://www.r-bloggers.com/circle-packing-with-r
  Slight modifications were made for code consistency and additional
  functionality.
}
