% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{Plot3DInteractive}
\alias{Plot3DInteractive}
\title{Plots 3D interactive}
\usage{
Plot3DInteractive(df = baea, x = "x", y = "y", z = "z")
}
\arguments{
\item{df}{Dataframe with x, y, z data, Default is 'baea'.}

\item{x}{Column name of x (longitude) locations.}

\item{y}{Column name of y (latitude) locations.}

\item{z}{Column name of z (elevation) locations}
}
\value{
3D plot from 'rgl'package
}
\description{
Makes a 3D plot with 'rgl' package
}
\details{
Can be any z, y, z, but labels are set for lat, long, elev data.
}
