% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{PlotBWRaster}
\alias{PlotBWRaster}
\title{Plot Rastet showing bandwidth}
\usage{
PlotBWRaster(in_raster, title = NA, color_pal = "Greens")
}
\arguments{
\item{in_raster}{RasterLayer}

\item{title}{character, title of plot}
}
\value{
ggplot
}
\description{
Plot Rastet showing bandwidth
}
