% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{PlotLocationCount}
\alias{PlotLocationCount}
\title{Plot location counts}
\usage{
PlotLocationsCount(df, id, individual, color_factor, start, end,
  breaks, tz, wrap)
}
\arguments{
\item{df}{Dataframe with id, datetime.}

\item{id}{Column name of unique identifier.}

\item{individual}{String, individual/s (from id column) to keep, format
should be c(id, id). Default is all.}

\item{color_factor}{Column names, factor to determine location point color.}

\item{start}{Date, filter start date. Default is 1970-01-01.}

\item{end}{Date, filter end date. Default is current date.}

\item{breaks}{Breaks on x-axis (i.e., "7 days"). Default is "14 days"}

\item{tz}{String, timezone. Default is "Etc/GMT+5".}

\item{wrap}{Logical, whether to wrap ribbon of panels into 2d. Default TRUE.}
}
\value{
Daily locations over a range of dates
}
\description{
Plots a linegraph of daily locations
}
\details{
Using color_factor on "behavior"
}
