% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{PlotLocationSunriseSunset}
\alias{PlotLocationSunriseSunset}
\title{Plots locations related to sunrise and sunset}
\usage{
PlotLocationSunriseSunset(
  df = df,
  by = "id",
  individual = "",
  color_factor = NULL,
  pal = NULL,
  b_pal = "Set1",
  start = NULL,
  end = NULL,
  breaks = "1 days",
  tz = "Etc/GMT+5",
  addsolartimes = FALSE,
  wrap = TRUE
)
}
\arguments{
\item{df}{Dataframe with id, lat, long, datetime.}

\item{by}{Column name to use to split, analyze, and merge data. Default is
"id".}

\item{individual}{String, individual/s (from id column) to keep, format
should be c(id, id), default is all}

\item{color_factor}{Column names, factor to determine location point color.}

\item{pal}{String, color palette for CreateVarsColors(). Default is NULL.}

\item{b_pal}{String, color palette from RColorBrewer for CreateVarsColors(),
default is "Set1"}

\item{start}{Date, filter start date. Default is 1970-01-01.}

\item{end}{Date, filter end date. Default is current date.}

\item{breaks}{Breaks on x-axis, i.e., "7 days".}

\item{tz}{Timezome. Default "Etc/GMT+5".}

\item{addsolartimes}{Logical, whether to not to run the AddSolarTimes() on
data first. Default FALSE.}

\item{wrap}{Logical, whether or not to wrap ribbon of panels into 2d.
Default is TRUE.}
}
\value{
Plot of locations over a range of dates
}
\description{
Plots locations in relation to sunrise, sunset, and solarnoon.
}
\details{
Using color_factor on "behavior"
}
