% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{PlotLogisticRange}
\alias{PlotLogisticRange}
\title{Plot a Logisitic function (with probability on the y-axis)}
\usage{
PlotLogisticRange(beta0, beta1, min_x = -0.25, max_x = 1.25)
}
\arguments{
\item{beta0}{numeric, intercept used in function}

\item{beta1}{numeric, slope used in function}

\item{min_X}{numeric, minimum on x axis}

\item{max_X}{numeric, maximum on x axis}
}
\value{
ggplot
}
\description{
Plot a Logisitic function (with probability on the y-axis)
}
\examples{
PlotLogisticZeroOneRange(intercept = 5, beta = -10)
}
