% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{PlotLogisticRange2Betas}
\alias{PlotLogisticRange2Betas}
\title{Plot Logistic Function for two continous variables [0, 1]}
\usage{
PlotLogisticRange2Betas(beta0 = -5, beta1 = 5, beta2 = 20,
  main_only = FALSE)
}
\arguments{
\item{beta0}{numeric, intercept}

\item{beta1}{numeric, slope parameter}

\item{beta2}{numeric, slope parameter}

\item{main_only}{logical, whether to return only main plot. Default is FALSE.}
}
\value{
plotly plot
}
\description{
Plot Logistic Function for two continous variables [0, 1]
}
\details{
Requres a Mapbox token and plotly username and api_key.
}
