% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{PlotLogisticRange2Par}
\alias{PlotLogisticRange2Par}
\title{Plot Logistic Function for two continous variables}
\usage{
PlotLogisticRange2Par(beta0 = -5, beta1 = 5, beta2 = 20)
}
\arguments{
\item{beta0}{numeric, intercept}

\item{beta1}{numeric, slope parameter}

\item{beta2}{numeric, slope parameter}
}
\value{
plotly plot
}
\description{
Plot Logistic Function for two continous variables
}
