% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{PlotLogisticRange3Betas}
\alias{PlotLogisticRange3Betas}
\title{Plot Logistic Function for two continous variables [0, 1] (with a constant
    third)}
\usage{
PlotLogisticRange3Betas(
  beta0 = 17.5,
  beta1 = -5,
  beta2 = -10,
  beta3 = -20,
  pred_held = 1,
  pred_value = 0.25,
  main_only = FALSE
)
}
\arguments{
\item{beta0}{numeric, intercept}

\item{beta1}{numeric, slope parameter}

\item{beta2}{numeric, slope parameter}

\item{beta3}{numeric, slope parameter}

\item{pred_held}{integer, surface to hold constant}

\item{pred_value}{integer, value for constant surface}

\item{main_only}{logical, whether to return only main plot. Default is FALSE.}
}
\value{
plotly plot
}
\description{
Plot Logistic Function for two continous variables [0, 1] (with a constant
    third)
}
\details{
Requres a Mapbox token and plotly username and api_key.
}
