% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{PlotMatrix}
\alias{PlotMatrix}
\title{Plot a matrix, with arguments for labels and coordinates}
\usage{
PlotMatrix(
  mat,
  digits = 2,
  title = NULL,
  label = TRUE,
  all_coords = TRUE,
  label_size = 4
)
}
\arguments{
\item{mat}{matrix}

\item{digits}{integer, number of digits shown on map}

\item{title}{character, plot title}

\item{label}{logical, whether to label}

\item{all_coords}{logical, whether to show all coordinates on x/y axis}

\item{label_size}{numberic, label size}
}
\value{
ggplot
}
\description{
Plot a matrix, with arguments for labels and coordinates
}
