% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{PlotMatrixCenterRow2D}
\alias{PlotMatrixCenterRow2D}
\title{Plot a matrix's center row}
\usage{
PlotMatrixCenterRow2D(matrix_in, title = NULL)
}
\arguments{
\item{matrix_in}{matrix}

\item{title}{character, title for plot}
}
\value{
ggplot
}
\description{
Plot a matrix's center row
}
