% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{PlotMove}
\alias{PlotMove}
\title{Plot move object}
\usage{
PlotMove(move)
}
\arguments{
\item{move}{A 'move' object}
}
\value{
Plot of move object
}
\description{
Wrapper for plotting a 'move' object
}
\details{
Only work on move objects
}
