% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{PlotOvernightDistances}
\alias{PlotOvernightDistances}
\title{Plot overnight distances}
\usage{
PlotOvernightDistances(df, individual, ylim, point_size)
}
\arguments{
\item{df}{Dataframe of locations with "date", "last", "step_length"}

\item{individual}{String, individual/s (from id column) to keep, format
should be id. Default is 'all'.}

\item{ylim}{Numeric, y-axis limit}

\item{point_size}{Numeric, point size for time datapoints}
}
\value{
A scatterplot
}
\description{
Scatterplot of overnight distances. Non-sequential days are removed.
}
