% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{PlotOvernightDistancesECDF}
\alias{PlotOvernightDistancesECDF}
\title{Plot overnight distances empirical distribution function}
\usage{
PlotOvernightDistancesECDF(df = baea, id = NULL, xlim = NULL,
  line_size = 2)
}
\arguments{
\item{df}{Dataframe of locations with "date", "last", "step_length".}

\item{id}{Column name of unique identifier.}

\item{xlim}{Numeric, x-axis limit. Default is NULL.}

\item{line_size}{Numeric, point size for time datapoints.}
}
\value{
Scatterplot
}
\description{
Empirical cumulative distribution; non-sequential days removed.
}
\details{
Requires RemoveNonsequentialDays() function.
}
