% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{PlotRasterCenterCellDistances}
\alias{PlotRasterCenterCellDistances}
\title{Plot raster displaying distances from center cell}
\usage{
PlotRasterCenterCellDistances(max_r, cellsize, col)
}
\arguments{
\item{cellsize}{raster cell size}

\item{max_r}{maximum circular radius from center cell}

\item{col}{color palette for base raster, default is "Spectral" from the
'RColorBrewer' package. Must have same number of colors as the number of
breaks as the data (i.e., ceiling(max_r/cellsize)).}

\item{label_size}{numeric, cell label size, default is 4.}
}
\value{
plot of cells
}
\description{
Plot a raster showing the cell distances and max_r based on cell size
}
\examples{
PlotRasterCenterCellDistances(cellsize = 15, max_r = 90)
PlotRasterCenterCellDistances(cellsize = 15, max_r = 90, col = rainbow(6))
PlotRasterCenterCellDistances(cellsize = 15, max_r = 95, col = rainbow(7))

}
