% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{PlotRasterCenterRow2D}
\alias{PlotRasterCenterRow2D}
\title{Plot a raster's center row}
\usage{
PlotRasterCenterRow2D(raster_in, title = NULL)
}
\arguments{
\item{raster_in}{RasterLayer}

\item{title}{character, title of plot}
}
\value{
ggplot
}
\description{
Plot a raster's center row
}
