% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{PlotRasterKernSlice}
\alias{PlotRasterKernSlice}
\title{Wrapper function to create, slice, and plot a Gaussian Kernel Raster}
\usage{
PlotRasterKernSlice(raster_in = NULL, sigma = 1, nrow = 21,
  shift_n = 0, cell_size = 30, title = NULL, normalize = FALSE,
  dnorm_line = FALSE)
}
\arguments{
\item{raster_in}{Raster*, overrides sigma and nrow below when provided.
Default is NULL which then uses the parameters below to create a Raster.}

\item{sigma}{numeric, sigma (or sd) of distribution - scaled to cell_size}

\item{nrow}{numeric, number of rows/columns in RasterLayer}

\item{shift_n}{numberic, shift in lat/long origin position}

\item{cell_size}{numeric, cell size of RasterLayer}

\item{title}{character, main title. If not provided, a default is used.}

\item{normalize}{logical, whether or not to normalize values to sum to 1}

\item{dnorm_line}{logical, whether or not to add a normal distribution curve
to plot. The curve will be scaled to the cell_size and sigma provided.}
}
\value{
ggplot
}
\description{
Wrapper function to create, slice, and plot a Gaussian Kernel Raster
}
