% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{PrintRasterNames}
\alias{PrintRasterNames}
\title{Prints the position and name of the rasters in a RasterStack or RasterBrick}
\usage{
PrintRasterNames(raster)
}
\arguments{
\item{raster}{RasterStack or RasterBrick}
}
\value{
Prints a list of positions and names
}
\description{
Prints the position and name of the rasters in a RasterStack or RasterBrick
}
