% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{Rad2Deg}
\alias{Rad2Deg}
\title{Converts radians to degrees}
\usage{
Rad2Deg(radians)
}
\arguments{
\item{radians}{numeric, angle in radians}
}
\value{
Angle in degrees
}
\description{
Converts radians to degrees
}
