% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{RasterizeOsMDownload}
\alias{RasterizeOsMDownload}
\title{Convert basemaps download from 'OpenStreetMaps' to Raster for use in tm_map}
\usage{
RasterizeOsMDownload(osm_download)
}
\arguments{
\item{osm_download}{download from OpenStreetMaps::openmap()}
}
\value{
a RasterLayer
}
\description{
Convert basemaps download from 'OpenStreetMaps' to Raster for use in tm_map
}
