% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{RasterizeROSMDownload}
\alias{RasterizeROSMDownload}
\title{Convert basemaps download from 'rosm' to Raster for use in tm_map}
\usage{
RasterizeROSMDownload(rosm_download)
}
\arguments{
\item{rosm_download}{download from rosm::osm.raster()}
}
\value{
a RasterLayer
}
\description{
Convert basemaps download from 'rosm' to Raster for use in tm_map
}
