% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{RemoveNonsequentialDays}
\alias{RemoveNonsequentialDays}
\title{Remove non-sequential days}
\usage{
RemoveNonsequentialDays(df)
}
\arguments{
\item{df}{Dataframe of location, default is "baea".}
}
\value{
Dataframe
}
\description{
Removes days that do not have GPS data for the next day
}
\details{
used in PlotOvernightDistancesECD().
}
