% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{RescaleAbsoluteDistanceRaster}
\alias{RescaleAbsoluteDistanceRaster}
\title{Rescale binary Raster to absolute distance Raster (0 at most 'internal' cell,
    highest value at most 'distant' cell)}
\usage{
RescaleAndUniformRaster(raster)
}
\arguments{
\item{raster}{RasterLayer}
}
\value{
RasterLayer
}
\description{
Rescale binary Raster to absolute distance Raster (0 at most 'internal' cell,
    highest value at most 'distant' cell)
}
