% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{RescaleAndUniformRaster}
\alias{RescaleAndUniformRaster}
\title{Rescale raster from 0-1 and redistribute values to a uniform distribution}
\usage{
RescaleAndUniformRaster(raster)
}
\arguments{
\item{raster}{RasterLayer}
}
\value{
RasterLayer
}
\description{
Rescale raster from 0-1 and redistribute values to a uniform distribution
}
