% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{RotateRaster}
\alias{RotateRaster}
\title{Rotates a Raster in 2-dimensional space}
\usage{
RotateRaster(raster, angle, resolution)
}
\arguments{
\item{raster}{Raster}

\item{angle}{numeric, degrees to rotate Raster}

\item{resolution}{numeric, resolution of output Raster}
}
\value{
A raster
}
\description{
Rotates a Raster in 2-dimensional space
}
\examples{
library(raster)
x <- raster(matrix(1:(15*25), nrow = 15), xmn = -1000, xmx = 1000,
ymn = -1000, ymx = 1000)
plot(x, main="Original")
plot(RotateRaster(x, 30, 10), main = paste("Rotated by 30 degrees"))
plot(RotateRaster(x, 75, 10), main = paste("Rotated by 75 degrees"))
plot(RotateRaster(x, 180, 10), main = paste("Rotated by 180 degrees"))
plot(RotateRaster(x, 300, 10), main = paste("Rotated by 300 degrees"))

}
