% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{SendWeeklyData}
\alias{SendWeeklyData}
\title{Creates and emails weekly .kml data}
\usage{
SendWeeklyData(data = data, send_email = TRUE, date = now(),
  to = c("erynn.call@maine.gov", "charlie.todd@maine.gov"))
}
\arguments{
\item{data}{Dataframe of deployed data.}

\item{send_email}{Logical, whether or not to send an email, default is TRUE.}

\item{date}{Date within week to update, default is now().}

\item{to}{String, recipients of email, default: c("erynn.call@maine.gov",
"charlie.todd@maine.gov").}
}
\value{
Creates files in "C:/Work/R/Data/BAEA/Telemetry" and
}
\description{
Creates weekly .kml file, and sends an email with that .kml file.
}
\details{
Specific to my Bald Eagle Project.
}
