% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_style_map.R
\name{si_style_map}
\alias{si_style_map}
\title{Standard SI Style for Thematic Maps}
\usage{
si_style_map(...)
}
\arguments{
\item{...}{Any other ggplot theme arguments}
}
\description{
This style borrows from Timo Grossenbacher's thematic maps to match OHA/SIEI Viz Products
}
\examples{
\dontrun{
get_admin0(countries = "Kenya") \%>\%
  ggplot() +
  geom_sf(fill = glitr::grey10k) +
  labs(
    title = "COUNTRY ABC - Sample thematic map",
    subtitle = "Region XYZ is over testing your males",
    caption = paste0("OHA/SIEI - Data from DATIM FY?? Q?, ", Sys.Date())
  ) +
  si_style_map()
}

}
