% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{SmoothRaster}
\alias{SmoothRaster}
\title{Smooth a RasterLayer using 'smoothie' package
Wrapper function to convert a Raster to a matrix, smooth the matrix using
    gauss2dsmooth(), and convert the smoothed matrix back into a Raster}
\usage{
SmoothRaster(sigma = sigma, covar = covar)
}
\arguments{
\item{sigma}{numeric, bandwidth (in cells) for gauss2dsmooth}

\item{covar}{RasterLayer name in quotes (e.g., "covar1")}
}
\value{
RasterLayer
}
\description{
Smooth a RasterLayer using 'smoothie' package
Wrapper function to convert a Raster to a matrix, smooth the matrix using
    gauss2dsmooth(), and convert the smoothed matrix back into a Raster
}
