% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{spdf_points}
\alias{spdf_points}
\title{Generate Point Spatial DataFrame}
\usage{
spdf_points(.data, lat = "latitude", long = "longitude", crs = 4326)
}
\arguments{
\item{.data}{Location data as a data frame, use \code{extract_facilities}}

\item{lat}{Column name for latitude, default value is latitude}

\item{long}{Column name for longitude, default value is longitude}

\item{crs}{Coordinate Reference System, default value is EPSG Code for WGS 1984}
}
\value{
list of spatial data frames
}
\description{
Generate Point Spatial DataFrame
}
\examples{
\dontrun{

cntry <- "Ethiopia"
level_fac <- get_ouorglevel(operatingunit = cntry, org_type = "facility")
df_facs <- extract_locations(country = cntry, level = level_fac)
df_facs <- df_facs \%>\% extract_facilities()
df_locs <- df_facs \%>\% select(-c(geom_type:nested))

spdf <- spdf_points(.data = df_locs)
}


}
