% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_distributions.R
\name{spdist_ovc_hivstat_pos}
\alias{spdist_ovc_hivstat_pos}
\title{Map the distribution of OVC_HIVSTAT_POS PSNU Results}
\usage{
spdist_ovc_hivstat_pos(country, fy, df_psnu, geo_psnu, terr_path = NULL)
}
\arguments{
\item{country}{Operating Unit name}

\item{fy}{Fiscal Year (4 digit)}

\item{df_psnu}{PSNU by IM Dataset (data frame)}

\item{geo_psnu}{PSNU Geodata (sf data frame)}

\item{terr_path}{Path to terrain raster file}
}
\description{
Map the distribution of OVC_HIVSTAT_POS PSNU Results
}
\examples{
\dontrun{
spdist_ovc_hivstat_pos(country = "saturn", fy = 3030,
    df_psnu = psnu_by_im, geo_psnu = saturn_districts)
spdist_ovc_hivstat_pos(country = "saturn", fy = 3030,
    df_psnu = psnu_by_im,
    geo_psnu = saturn_districts,
    terr_path = path_to_terr)
}

}
