% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{SummarizeDailyLocations}
\alias{SummarizeDailyLocations}
\title{Summarize daily locations}
\usage{
SummarizeDailyLocations(df = df)
}
\arguments{
\item{df}{Dataframe of locations, needs "id" and "date" columns}
}
\value{
Dataframe of total daily location, including 0 for days w/o data.
}
\description{
Gets the daily number of locations for a location dataset
}
\details{
Used in PlotDailyLocationsCount()
}
