% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{SummarizeLocations}
\alias{SummarizeLocations}
\title{Summarize locations}
\usage{
SummarizeLocations(df)
}
\arguments{
\item{df}{Dataframe with location data}

\item{pdf}{Logical, whether or not to write PDF. Default is FALSE.}
}
\value{
Table and PDF (optional)
}
\description{
Calculates summarize stats for altitude, speed, moving and total locations
}
\details{
default outputs are specific to my file directories and locations
}
