% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kml.R
\name{ZipKML}
\alias{ZipKML}
\title{ZipKML}
\usage{
ZipKML(kml, image)
}
\arguments{
\item{kml}{filename for RasterStack or RasterBrick}

\item{image}{filename for image file}
}
\value{
writes a zip file from a kml
}
\description{
Zip a kml file
}
\details{
Based on function from 'raster' package
}
