% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admins_map.R
\name{admins_map}
\alias{admins_map}
\title{Generate admin map}
\usage{
admins_map(countries, add_neighbors = FALSE)
}
\arguments{
\item{countries}{list of countries to map}

\item{add_neighbors}{should the map include the neighbor countries}
}
\value{

}
\description{
Generate admin map
}
\examples{
\dontrun{
terrain_map(countries = list("Zambia"))
terrain_map(countries = list("Zambia"), add_neighbors = TRUE)
}

}
