% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{CalculateTerrainMetric}
\alias{CalculateTerrainMetric}
\title{Calculate terrain metrics on elevation RasterLayer}
\usage{
CalculateTerrainMetric(elev, size, metric)
}
\arguments{
\item{elev}{RasterLayer, elevation data raster}

\item{size}{numeric, dimensions of analysis window (nrow and ncol). Must be
odd number. When size == 0, it is internally converted to size == 1 which
is necessary for the metric to be calculated (and not just revert to the
'elev' layer. This was done to allow an optimization procedure to include
a "sigma = 0", even though that would be identical to size == 1.}

\item{metric}{character, vector containing one of these options: tpi, tri,
roughness (see Details)}
}
\value{
RasterLayer
}
\description{
Calculate terrain metrics on elevation RasterLayer
}
\details{
metrics based on examples given in terrain() in 'raster' package
}
