% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{cntry_polygons}
\alias{cntry_polygons}
\title{Extract Country Polygons}
\usage{
cntry_polygons(spdf, cntry, attrs)
}
\arguments{
\item{spdf}{VcPolygons data as Spatial Data Frame}

\item{cntry}{Country name}

\item{attrs}{Country orgunits, output of \code{grabr::datim_orgunits(cntry, reshape = TRUE)}}
}
\value{
list of spatial data frames
}
\description{
Extract Country Polygons
}
\examples{
\dontrun{

cntry = "Nigeria"

spdf <- gisr::get_vcpolygons(path = glamr::si_path("path_vector"), name = "VcPepfarPolygons.shp")

df_attrs <- grabr::get_attributes(cntry)

cntry_polygons(spdf = spdf, cntry = "Zambia", attrs = df_attrs)

}

}
