% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{dview}
\alias{dview}
\title{View attributes from simple feature object}
\usage{
dview(geodata, console = FALSE)
}
\arguments{
\item{geodata}{spatial data}

\item{console}{view in console? default false}
}
\description{
View attributes from simple feature object
}
\examples{
\dontrun{
 library(gisr)

 adm0 <- get_admin0("Ghana")

 adm0 \%>\% dview(console = TRUE)
}

}
