% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{export_spdf}
\alias{export_spdf}
\title{Save shapefile}
\usage{
export_spdf(spdf, name)
}
\arguments{
\item{spdf}{sf object}

\item{name}{filename with full path}
}
\value{
\preformatted{boolean
}
}
\description{
Save shapefile
}
\examples{
\dontrun{
 library(gisr)
 library(sf)

 shp <- get_admin0(countries = "Nigeria")

 export_spdf(spdf = shp, name = "./GIS/nga_country_boundaries")
 export_spdf(spdf = shp, name = "./GIS/nga_country_boundaries.shp")
}

}
