% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_attributes}
\alias{extract_attributes}
\title{Extract Orgunit Boundaries Attributes}
\usage{
extract_attributes(country, folderpath = NULL)
}
\arguments{
\item{country}{OU/country}

\item{folderpath}{Local directory of files}
}
\value{
df
}
\description{
Extract Orgunit Boundaries Attributes
}
\examples{
\dontrun{
 library(tidyverse)
 library(gisr)

 extract_attributes(country = "Nigeria")
 extract_attributes(country = "Nigeria", folderpath = glamr::si_path("path_vector"))
}

}
