% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_locations.R
\name{extract_facilities}
\alias{extract_facilities}
\title{Extract facility sites}
\usage{
extract_facilities(.data, sites = NULL)
}
\arguments{
\item{.data}{Datim organisation units data frame}

\item{sites}{Data Frame of MER Sites by IM with Results and/or Targets [cols: orgunituid, sitename]}
}
\description{
Extract facility sites
}
\examples{
\dontrun{
extract_facilities(df_orgunits)
df_orgunits \%>\% extract_facilities()
}

}
