% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_locations.R
\name{extract_facilities}
\alias{extract_facilities}
\title{Extract facility sites}
\usage{
extract_facilities(.data, targets = NULL)
}
\arguments{
\item{.data}{Datim organisation units data frame}

\item{targets}{Data Frame of MER Results / Targets}
}
\description{
Extract facility sites
}
\examples{
\dontrun{
extract_facilities(df_orgunits)
df_orgunits \%>\% extract_facilities()
}

}
