% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_locations.R
\name{extract_locations}
\alias{extract_locations}
\title{Extract location data}
\usage{
extract_locations(
  country,
  username,
  password,
  level = NULL,
  add_geom = TRUE,
<<<<<<< HEAD
  baseurl = NULL
=======
  base_url = NULL
>>>>>>> main
)
}
\arguments{
\item{country}{PEPFAR Operating Unit or Country name}

\item{username}{Datim Account Username}

\item{password}{Datim Account password}

\item{level}{PEPFAR Org Level, optional}

\item{add_geom}{Include geometry column, default value is true}

\item{base_url}{Datim URL}
}
\value{
A dataframe or Null if not match
}
\description{
Extract location data
}
\examples{
\dontrun{

extract_locations("<saturn>", "<my_username>", "<my_password>")
}

}
