% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_locations.R
\name{extract_locations}
\alias{extract_locations}
\title{Extract location data}
\usage{
extract_locations(
  country,
  level = NULL,
  add_geom = TRUE,
  username = NULL,
  password = NULL,
  base_url = NULL
)
}
\arguments{
\item{country}{PEPFAR Operatingunit}

\item{level}{PEPFAR Org Level}

\item{add_geom}{Include geometry column}

\item{username}{Datim Account Username}

\item{password}{Datim Account password}

\item{base_url}{Datim URL}
}
\description{
Extract location data
}
\examples{
\dontrun{
extract_locations("saturn", "my_username", "my_password")
}
}
