% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{geo_neighbors}
\alias{geo_neighbors}
\title{Get neighbors of a given contry}
\usage{
geo_neighbors(countries, scale = "large", crs = 4326)
}
\arguments{
\item{countries}{a list of country names}

\item{scale}{spatial resolution of the geodata}

\item{crs}{coordinates reference system}
}
\value{
simple feature class
}
\description{
Get neighbors of a given contry
}
\examples{
\dontrun{
geo_neighors(countries = list("Zambia"))
geo_neighors(countries = list("Zambia", "Malawi"))
}

}
