% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{geo_neighbors}
\alias{geo_neighbors}
\title{Get neighbors of a given contry}
\usage{
geo_neighbors(countries, scale = "large", crs = 4326, crop = FALSE)
}
\arguments{
\item{countries}{Country or list of countries names}

\item{scale}{Spatial resolution of the geodata}

\item{crs}{Coordinates reference system, default is WGS84 (EPGS:4326)}

\item{crop}{Crop sfc to focus countries extent?}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      simple feature class
}\if{html}{\out{</div>}}
}
\description{
Get neighbors of a given contry
}
\examples{
\dontrun{
 library(gisr)

 cntry <- "Zambia"

 # Country + neighbors boundaries
 cntries <- geo_neighbors(countries = cntry)

 # Country + neighbors boundaries: crop to country extent
 cntries <- geo_neighbors(countries = cntry, crop = TRUE)

 cntries \%>\%
   dplyr::select(name) \%>\%
   attributes()
}

}
