% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adm_boundaries.R
\name{get_adm_boundaries}
\alias{get_adm_boundaries}
\title{Get geospatial data from GADM}
\usage{
get_adm_boundaries(country_code, adm_level = 0, download = TRUE, geo_path = "")
}
\arguments{
\item{country_code}{ISO3 code}

\item{adm_level}{country administrative boundaries level}

\item{geo_path}{where to save / read data from}

\item{dowload}{Save a copy}
}
\value{
geodata as sf with crs = 4326
}
\description{
This function extracts / reads geospatial data
for specified country and administractive boundary levels
}
\examples{
\dontrun{
get_adm_boundaries("UGA")
get_adm_boundaries("UGA", adm_level=2, geo_path="GIS")
}

}
