% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{get_admin0}
\alias{get_admin0}
\title{Get admin level 0 boundary}
\usage{
get_admin0(countries, scale = c("medium", "large", "small"), crs = 4326)
}
\arguments{
\item{countries}{list of country names}

\item{scale}{spatial resolution of the geodata}

\item{crs}{coordinates reference system}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     simple feature class
}\if{html}{\out{</div>}}
}
\description{
sf boundaries data for a given country
}
\examples{
\dontrun{
  library(gisr)

  get_admin0(countries = list("Zambia"))
}

}
