% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_attributes}
\alias{get_attributes}
\title{Get Attributes Data for Orgunit Boundaries}
\usage{
get_attributes(country, username = NULL, password = NULL, folderpath = NULL)
}
\arguments{
\item{country}{OU/country}

\item{username}{Datim username}

\item{password}{Datim password}

\item{folderpath}{Local directory of files}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      OU Orgunit level as df
}\if{html}{\out{</div>}}
}
\description{
Get Attributes Data for Orgunit Boundaries
}
\note{
This will attempt to read data from local directory when folderpath is not set to null.
If null, username and password will be required
}
\examples{
\dontrun{
 library(gisr)

 get_attributes(country = "Nigeria")
}

}
