% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_map.R
\name{get_basemap}
\alias{get_basemap}
\title{Get Basemap}
\usage{
get_basemap(spdf, terr, country = NULL, add_admins = FALSE)
}
\arguments{
\item{spdf}{PEPFAR ORGs Spatial Data}

\item{terr}{RasterLayer}

\item{country}{OU or Country Name}

\item{add_admins}{Should the sub-admins be added? Default is false}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       ggplot plot of base map
}\if{html}{\out{</div>}}
}
\description{
Get Basemap
}
\examples{
\dontrun{
library(gisr)
library(sf)

 shp <- get_pepfar_shp(shp_path = glamr::si_path("path_vector"), add_attr = TRUE)
 ras <- get_raster(terr_path = glamr::si_path("path_raster"))

 get_basemap(spdf = shp, country = "Nigeria", terr = ras)
}

}
