% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hexbins.R
\name{get_grids}
\alias{get_grids}
\title{Create square grids}
\usage{
get_grids(spdf, size = 15000)
}
\arguments{
\item{spdf}{input spatial data frame}

\item{size}{size of each hex bin in meters, default set to 15K meters (15KM)}
}
\value{
country hex polygon as feature class
}
\description{
Create square grids
}
\examples{
\dontrun{
 library(gisr)

 shp <- get_admin0(countries = "Nigeria")

 get_grids(shp, 10000)
}

}
