% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hexbins.R
\name{get_hexbins}
\alias{get_hexbins}
\title{Create an hexbins polygon feature class}
\usage{
get_hexbins(country_code, adm_level = 0, geo_path = NULL, size = NULL)
}
\arguments{
\item{country_code}{iso3 code}

\item{geo_path}{path to geodata}

\item{size}{size of each hex bin in meters}

\item{level}{country administrative unit level}
}
\value{
country hex polygon as feature class
}
\description{
Create an hexbins polygon feature class
}
