% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_levels}
\alias{get_levels}
\title{Get all orgunits levels in org hierarchy}
\usage{
get_levels(
  username = NULL,
  password = NULL,
  baseurl = "https://final.datim.org/"
)
}
\arguments{
\item{username}{DATIM username, recommed using \code{datim_user()}}

\item{password}{DATIM password, recommend using \code{datim_pwd()}}

\item{baseurl}{base API url, default = https://final.datim.org/}
}
\value{
df
}
\description{
Get all orgunits levels in org hierarchy
}
\note{
Same as \code{glamr::identify_levels()} or \code{Wavelength::identify_levels()}
}
\examples{
\dontrun{
  get_levels() }

}
