% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_levels}
\alias{get_levels}
\title{Get all orgunits levels in org hierarchy}
\usage{
get_levels(username, password, reshape = FALSE, baseurl = NULL)
}
\arguments{
\item{username}{DATIM username, recommed using \code{datim_user()}}

\item{password}{DATIM password, recommend using \code{datim_pwd()}}

\item{reshape}{Reshape results to long format}

\item{baseurl}{base API url, default = https://final.datim.org/}
}
\value{
df
}
\description{
Get all orgunits levels in org hierarchy
}
\note{
Same as \code{glamr::identify_levels()} or \code{Wavelength::identify_levels()}
}
\examples{
\dontrun{
  library(gisr)

  # Get PEPFAR Org Levels
  get_levels()
 }

}
