% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_ouorguids}
\alias{get_ouorguids}
\title{Get OU Org UIDS}
\usage{
get_ouorguids(
  ouuid,
  level = 4,
  username = NULL,
  password = NULL,
  baseurl = "https://final.datim.org/"
)
}
\arguments{
\item{ouuid}{OU uid}

\item{level}{org level}

\item{username}{DATIM Username}

\item{password}{DATIM password, recommend using `mypwd()`}

\item{baseurl}{base url for the API, default = https://final.datim.org/}
}
\value{
ORG UIDS as tibble
}
\description{
Get OU Org UIDS
}
\note{
Use `get_orguids()` for levels above 4
}
\examples{
\dontrun{
 orgs <- get_ouorguids(ouuid = "<ou-uid-goes-here>", level = 4)}

}
