% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_ouuid}
\alias{get_ouuid}
\title{Get Operatingunit / Country Org UID}
\usage{
get_ouuid(operatingunit, username = NULL, password = NULL)
}
\arguments{
\item{operatingunit}{Operatingunit name}

\item{username}{Datim Account username, recommend using \code{datim_user()}}

\item{password}{Datim Account Password, recommend using \code{datim_pwd()}}
}
\value{
uid
}
\description{
Get Operatingunit / Country Org UID
}
\examples{
\dontrun{
  library(gisr)

  # get orgunit for specific OU/Country: kenya
  get_ouuid(operatingunit = "Kenya")
}

}
