% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_map.R
\name{get_raster}
\alias{get_raster}
\title{Get Terrain Raster dataset}
\usage{
get_raster(folderpath, name = NULL, rename = FALSE, ...)
}
\arguments{
\item{folderpath}{Path to raster file, default will be \code{si_path('path_raster')}}

\item{name}{Name of the raster file (with extension), default is set to terrain raster \code{SR_LR.tif}}

\item{rename}{Should the RasterLayer be renamed? If yes, the name is changed to \code{value}}

\item{...}{Additional arguments to be passed to \code{base::list.files}. Eg: \verb{Use ignore.case = TRUE} for non case sensitive search}
}
\value{
RasterLayer
}
\description{
Get Terrain Raster dataset
}
\examples{
\dontrun{
 library(glamr)
 library(gisr)

 get_raster()

 get_raster(name = "sample.tif")
 get_raster(path = "./geodata/raster", name = "sample.tif")
}

}
