% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_map.R
\name{get_raster}
\alias{get_raster}
\title{Get Terrain Raster dataset}
\usage{
get_raster(path = NULL, name = NULL, ...)
}
\arguments{
\item{path}{Path to raster file, default will be \code{si_path('path_raster')}}

\item{name}{Name of the raster file (with extension), default is set to terrain raster \code{SR_LR.tif}}

\item{...}{Additional arguments to be passed to \code{base::list.files}. Eg: \verb{Use ignore.case = TRUE} for non case sensitive search}
}
\value{
RasterLayer
}
\description{
Get Terrain Raster dataset
}
\examples{
\dontrun{
get_raster(terr_path = glamr::si_path("path_raster"))
}

}
