% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{get_terrain}
\alias{get_terrain}
\title{Get terrain data for an AOI (Countries)}
\usage{
get_terrain(
  countries = list("Zambia"),
  mask = FALSE,
  buffer = 0.2,
  terr_path = NULL
)
}
\arguments{
\item{countries}{list of the country names}

\item{mask}{should the extracted data match the exact bountries limits}

\item{buffer}{extend AOI extent by x}

\item{terr_path}{path to terrain raster file}
}
\value{
spdf spatial dataframe
}
\description{
Get terrain data for an AOI (Countries)
}
\examples{
\dontrun{
get_terrain(countries = list("Zambia"))
get_terrain(countries = list("Zambia"), mask = TRUE)
get_terrain(countries = list("Zambia"), buffer = .5, terr_path = "../../HDX_Data")
}
}
