% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_map.R
\name{get_terrain}
\alias{get_terrain}
\title{Get terrain data for an AOI (Countries)}
\usage{
get_terrain(
  countries = list("Zambia"),
  mask = FALSE,
  buffer = 0.1,
  terr = NULL
)
}
\arguments{
\item{countries}{List of the country names or sf object}

\item{mask}{Should the extracted data match the exact boundary limits?}

\item{buffer}{Extend AOI extent by x}

\item{terr}{RasterLayer or Path to terrain raster file}
}
\value{
\preformatted{      spdf spatial dataframe
}
}
\description{
Get terrain data for an AOI (Countries)
}
\examples{
\dontrun{
 library(gisr)
 library(sf)

 get_terrain(countries = list("Zambia"))
 get_terrain(countries = list("Zambia"), mask = TRUE)
 get_terrain(countries = list("Zambia"), buffer = .5, terr = "../../HDX_Data")

}
}
