% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{get_vcpolygons}
\alias{get_vcpolygons}
\title{Get PEPFAR Visual Crossing Polygons}
\usage{
get_vcpolygons(folderpath, name = NULL)
}
\arguments{
\item{folderpath}{Path to PEPFAR Global Shapefile}

\item{name}{Name or pattern of shapefile}
}
\value{
sf object
}
\description{
Get PEPFAR Visual Crossing Polygons
}
\examples{
\dontrun{
 library(gisr)

 shp_pepfar <- get_vcpolygons()
 shp_pepfar <- get_vcpolygons(path = "./GIS")
 shp_pepfar <- get_vcpolygons(path = "./GIS", name = "VcPepfarPolygons.shp")
}

}
