% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{gview}
\alias{gview}
\title{Plot sf features}
\usage{
gview(geodata, ...)
}
\arguments{
\item{geodata}{spatial data as sf object}

\item{...}{arguments passed to geom_sf}
}
\description{
Plot sf features
}
\examples{
\dontrun{
 library(gisr)

 adm0 <- get_admin0("Ghana")

 adm0 \%>\%
  dplyr::select(name) \%>\%
  gview()
}

}
