% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_map.R
\name{terrain_map}
\alias{terrain_map}
\title{Generate a terrain map}
\usage{
terrain_map(countries, terr_path = NULL, add_neighbors = FALSE)
}
\arguments{
\item{countries}{list of countries to map}

\item{terr_path}{path for terrain raster file}

\item{add_neighbors}{should the map include the neighbor countries}
}
\value{

}
\description{
Generate a terrain map
}
\examples{
\dontrun{
terrain_map(countries = list("Zambia"))
terrain_map(countries = list("Zambia"), add_neighbors = TRUE)
}

}
