% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_map.R
\name{terrain_map}
\alias{terrain_map}
\title{Generate a terrain map}
\usage{
terrain_map(
  countries,
  adm0 = NULL,
  adm1 = NULL,
  add_neighbors = FALSE,
  add_labels = FALSE,
  mask = FALSE,
  terr = NULL
)
}
\arguments{
\item{countries}{List of countries to map}

\item{adm0}{Admin 0 boundaries, optional sf geodata}

\item{adm1}{Admin 1 boundaries, optional sf geodata}

\item{add_neighbors}{Should the map include the neighbor countries}

\item{add_labels}{Add neighbors countries's names (works only when add_neighbors is TRUE)}

\item{mask}{Should the extracted data match the exact boundary limits?}

\item{terr}{RasterLayer dataset or Path for terrain raster file}
}
\value{

}
\description{
Generate a terrain map
}
\examples{
\dontrun{
library(gisr)
terrain_map(countries = list("Zambia"))
terrain_map(countries = list("Zambia"), add_neighbors = TRUE)
}

}
