% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kml.R
\name{UpdateIndByYearKMLs}
\alias{UpdateIndByYearKMLs}
\title{Creates or updates .kml files of each individual by year}
\usage{
UpdateIndByYearKMLs(data, update_only = TRUE, update_gdrive = TRUE)
}
\arguments{
\item{data}{Dataframe of location data.}

\item{update_only}{Logical, whether or not to just update current year's
data. Default is TRUE.}

\item{update_gdrive}{Logical, whether to update the files on the GDrive.
Default is TRUE.}
}
\value{
Creates .kml files in "Data/GPS" and copies them to
  "Google Drive/BAEA Project/Telemetry Data/Individuals" folders.
}
\description{
Creates and updates year KML files on local and GDrive folders.
}
