% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{zip_shapefiles}
\alias{zip_shapefiles}
\title{Compress all shapefile components into a zipped file}
\usage{
zip_shapefiles(filename, dest_folder = NULL)
}
\arguments{
\item{filename}{Shapefile full name}

\item{dest_folder}{Where to place the zipped files}
}
\value{
\preformatted{       Boolean
}
}
\description{
Compress all shapefile components into a zipped file
}
\examples{
\dontrun{
 library(gisr)
 library(sf)

 shp <- get_admin0(countries = "Nigeria")

 fname <- ./GIS/nga_country_boundaries"

 export_spdf(spdf = shp, name = fname)

 zip_shapefiles(filename = fname, dest_folder = "./GIS/)
}

}
