% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{zip_shapefiles}
\alias{zip_shapefiles}
\title{Compress all shapefile components into a zipped file}
\usage{
zip_shapefiles(filename, folderpath = NULL)
}
\arguments{
\item{filename}{Shapefile full path and name}

\item{folderpath}{Where to place the zipped files}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       Boolean
}\if{html}{\out{</div>}}
}
\description{
Compress all shapefile components into a zipped file
}
